Add county name, FIPs, and geometry to data frame with zip codes

Add county name, FIPs, and geometry to data frame with zip codes

Usage

add_counties(
  df,
  statecol,
  zipcol,
  add_fips = FALSE,
  add_geo = FALSE,
  as_sf = FALSE
)

Arguments

df

data frame

statecol

bare column name column containing state abbreviations

zipcol

bare column name containing zip codes

add_fips

add county FIPS code column, default is FALSE

add_geo

add county geometry column, default is FALSE

as_sf

convert tibble to an {sf} object, default is FALSE

Examples

if (FALSE) { # interactive()
nppes(npi = 1720098791) |>
      add_counties(statecol = state,
                   zipcol   = zip,
                   add_fips = TRUE,
                   add_geo  = TRUE,
                   as_sf    = TRUE)
}