Search the Medicare Monthly Enrollment API

Current information on the number of Medicare beneficiaries with hospital/medical coverage and prescription drug coverage, available for several geographical areas.

Usage

beneficiary_enrollment(
  year = NULL,
  month = NULL,
  level = NULL,
  state = NULL,
  state_name = NULL,
  county = NULL,
  fips = NULL,
  tidy = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

year

Calendar year of Medicare enrollment; current options are 2017 - 2022

month

Time frame of Medicare enrollment; options are Year or any month within the 12-month time span of the month in the data set's version name (listed here)

level

Geographic level of data; options are National, State, and County

state

Two-letter state abbreviation of beneficiary residence

state_name

Full state name of beneficiary residence

county

County of beneficiary residence

fips

FIPS code of beneficiary residence

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Details

The Medicare Monthly Enrollment data provides current monthly information on the number of Medicare beneficiaries with hospital/medical coverage and prescription drug coverage, available for several geographic areas including national, state/territory, and county. The hospital/medical coverage data can be broken down further by health care delivery (Original Medicare versus Medicare Advantage and Other Health Plans) and the prescription drug coverage data can be examined by those enrolled in stand-alone Prescription Drug Plans and those enrolled in Medicare Advantage Prescription Drug plans. The dataset includes enrollee counts on a rolling 12 month basis and also provides information on yearly trends. The dataset is based on information gathered from CMS administrative enrollment data for beneficiaries enrolled in the Medicare program available from the CMS Chronic Conditions Data Warehouse.

Links

Medicare Monthly Enrollment

Note

Update Frequency: Monthly

Examples

beneficiary_enrollment(year = NULL,
                       month = "Year",
                       level = "County",
                       state = "AL",
                       county = "Autauga")
#> # A tibble: 10 × 26
#>     year month level  state state_name county  fips  bene_total bene_orig
#>    <int> <chr> <chr>  <chr> <chr>      <chr>   <chr>      <int>     <int>
#>  1  2013 Year  County AL    Alabama    Autauga 01001       9323      6484
#>  2  2014 Year  County AL    Alabama    Autauga 01001       9589      6565
#>  3  2015 Year  County AL    Alabama    Autauga 01001       9938      6711
#>  4  2016 Year  County AL    Alabama    Autauga 01001      10214      6799
#>  5  2017 Year  County AL    Alabama    Autauga 01001      10510      5784
#>  6  2018 Year  County AL    Alabama    Autauga 01001      10645      5700
#>  7  2019 Year  County AL    Alabama    Autauga 01001      11059      5761
#>  8  2020 Year  County AL    Alabama    Autauga 01001      11251      5596
#>  9  2021 Year  County AL    Alabama    Autauga 01001      11396      5338
#> 10  2022 Year  County AL    Alabama    Autauga 01001      11578      5186
#> # ℹ 17 more variables: bene_ma_oth <int>, bene_aged_total <int>,
#> #   bene_aged_esrd <int>, bene_aged_no_esrd <int>, bene_dsb_total <int>,
#> #   bene_dsb_esrd <int>, bene_dsb_no_esrd <int>, bene_ab_total <int>,
#> #   bene_ab_orig <int>, bene_ab_ma_oth <int>, bene_rx_total <int>,
#> #   bene_rx_pdp <int>, bene_rx_mapd <int>, bene_rx_elig <int>,
#> #   bene_rx_full <int>, bene_rx_part <int>, bene_rx_none <int>

beneficiary_enrollment(year = 2021, level = "County", fips = "01001")
#> # A tibble: 13 × 26
#>     year month     level  state state_name county  fips  bene_total bene_orig
#>    <int> <chr>     <chr>  <chr> <chr>      <chr>   <chr>      <int>     <int>
#>  1  2021 Year      County AL    Alabama    Autauga 01001      11396      5338
#>  2  2021 January   County AL    Alabama    Autauga 01001      11371      5368
#>  3  2021 February  County AL    Alabama    Autauga 01001      11356      5355
#>  4  2021 March     County AL    Alabama    Autauga 01001      11362      5352
#>  5  2021 April     County AL    Alabama    Autauga 01001      11366      5329
#>  6  2021 May       County AL    Alabama    Autauga 01001      11376      5329
#>  7  2021 June      County AL    Alabama    Autauga 01001      11372      5332
#>  8  2021 July      County AL    Alabama    Autauga 01001      11399      5329
#>  9  2021 August    County AL    Alabama    Autauga 01001      11416      5322
#> 10  2021 September County AL    Alabama    Autauga 01001      11410      5329
#> 11  2021 October   County AL    Alabama    Autauga 01001      11425      5327
#> 12  2021 November  County AL    Alabama    Autauga 01001      11449      5338
#> 13  2021 December  County AL    Alabama    Autauga 01001      11455      5345
#> # ℹ 17 more variables: bene_ma_oth <int>, bene_aged_total <int>,
#> #   bene_aged_esrd <int>, bene_aged_no_esrd <int>, bene_dsb_total <int>,
#> #   bene_dsb_esrd <int>, bene_dsb_no_esrd <int>, bene_ab_total <int>,
#> #   bene_ab_orig <int>, bene_ab_ma_oth <int>, bene_rx_total <int>,
#> #   bene_rx_pdp <int>, bene_rx_mapd <int>, bene_rx_elig <int>,
#> #   bene_rx_full <int>, bene_rx_part <int>, bene_rx_none <int>

beneficiary_enrollment(year = 2022,
                       month = "July",
                       level = "State",
                       state_name = "Georgia")
#> # A tibble: 1 × 26
#>    year month level state state_name county fips  bene_total bene_orig
#>   <int> <chr> <chr> <chr> <chr>      <chr>  <chr>      <int>     <int>
#> 1  2022 July  State GA    Georgia    Total  13       1850422    915890
#> # ℹ 17 more variables: bene_ma_oth <int>, bene_aged_total <int>,
#> #   bene_aged_esrd <int>, bene_aged_no_esrd <int>, bene_dsb_total <int>,
#> #   bene_dsb_esrd <int>, bene_dsb_no_esrd <int>, bene_ab_total <int>,
#> #   bene_ab_orig <int>, bene_ab_ma_oth <int>, bene_rx_total <int>,
#> #   bene_rx_pdp <int>, bene_rx_mapd <int>, bene_rx_elig <int>,
#> #   bene_rx_full <int>, bene_rx_part <int>, bene_rx_none <int>

beneficiary_enrollment(level = "State", fips = "10")
#> # A tibble: 131 × 26
#>     year month     level state state_name county fips  bene_total bene_orig
#>    <int> <chr>     <chr> <chr> <chr>      <chr>  <chr>      <int>     <int>
#>  1  2013 Year      State DE    Delaware   Total  10        167686    155803
#>  2  2013 January   State DE    Delaware   Total  10        165415    153731
#>  3  2013 February  State DE    Delaware   Total  10        165732    153988
#>  4  2013 March     State DE    Delaware   Total  10        166144    154433
#>  5  2013 April     State DE    Delaware   Total  10        166543    154782
#>  6  2013 May       State DE    Delaware   Total  10        166860    155047
#>  7  2013 June      State DE    Delaware   Total  10        167294    155431
#>  8  2013 July      State DE    Delaware   Total  10        167851    155924
#>  9  2013 August    State DE    Delaware   Total  10        168302    156322
#> 10  2013 September State DE    Delaware   Total  10        168827    156808
#> # ℹ 121 more rows
#> # ℹ 17 more variables: bene_ma_oth <int>, bene_aged_total <int>,
#> #   bene_aged_esrd <int>, bene_aged_no_esrd <int>, bene_dsb_total <int>,
#> #   bene_dsb_esrd <int>, bene_dsb_no_esrd <int>, bene_ab_total <int>,
#> #   bene_ab_orig <int>, bene_ab_ma_oth <int>, bene_rx_total <int>,
#> #   bene_rx_pdp <int>, bene_rx_mapd <int>, bene_rx_elig <int>,
#> #   bene_rx_full <int>, bene_rx_part <int>, bene_rx_none <int>