Medicare Part B Utilization Statistics by Geography and Service

by_geography() allows you to search for information on services and procedures provided to Original Medicare (fee-for-service) Part B beneficiaries by physicians and other healthcare professionals, aggregated by geography and service.

The Geography and Service subset contains information on utilization, allowed amount, Medicare payment, and submitted charges organized nationally and state-wide by HCPCS code and place of service.

Links

Physician & Other Practitioners: by Geography and Service API

Update Frequency: Annually

Usage

by_geography(
  year,
  state = NULL,
  hcpcs_code = NULL,
  pos = NULL,
  level = NULL,
  fips = NULL,
  hcpcs_desc = NULL,
  drug = NULL,
  tidy = TRUE
)

Arguments

year

int (required); Year in YYYY format. Run helper function prac_years() to return a vector of the years currently available.

state

State where the provider is located, as reported in NPPES. Values include the 50 United States, District of Columbia, U.S. territories, Armed Forces areas, Unknown and Foreign Country. Data aggregated at the National level are identified by the word 'National'.

hcpcs_code

HCPCS code used to identify the specific medical service furnished by the provider.

pos

Identifies whether the place of service submitted on claims is a facility ("F") or non-facility ("O"). Non-facility is generally an office setting.

level

Level of geography by which the data will be aggregated. State indicates the data is aggregated to a single state identified in the Rendering Provider State column. National indicates the data is aggregated across all states for a given HCPCS Code Level.

fips

FIPS code of the referring provider state. This variable is blank when reported at the national level.

hcpcs_desc

Description of the HCPCS code for the specific medical service furnished by the provider.

drug

Flag that identifies whether the HCPCS code for the specific service furnished by the provider is a HCPCS listed on the Medicare Part B Drug Average Sales Price (ASP) File.

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Examples

if (FALSE) { # interactive()
by_geography(year = 2020, hcpcs_code = "0002A")
}