Search the Medicare Physician & Other Practitioners API by Provider and Service

Information on services and procedures provided to Original Medicare (fee-for-service) Part B (Medical Insurance) beneficiaries by physicians and other healthcare professionals; aggregated by provider and service.

Usage

by_service(
  year,
  npi = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par = NULL,
  hcpcs_code = NULL,
  drug = NULL,
  pos = NULL,
  tidy = TRUE
)

Arguments

year

integer (required); Year in YYYY format. Run helper function prac_years() to return a vector of the years currently available.

npi

National Provider Identifier for the rendering provider on the claim

first, last

Individual provider's name

organization

Organization name

credential

Individual provider's credentials

gender

Individual provider's gender; "F" (Female), "M" (Male)

entype

Provider entity type; "I" (Individual), "O" (Organization)

city

City where provider is located

state

State where provider is located

zip

Provider’s zip code

fips

Provider's state FIPS code

ruca

Rural-Urban Commuting Area Code (RUCA). Choices are:

Metro Area Core

"1": Primary flow within Urbanized Area (UA)

"1.1": Secondary flow 30-50% to larger UA

Metro Area High Commuting

"2": Primary flow 30% or more to UA

"2.1": Secondary flow 30-50% to larger UA - Metro Area Low Commuting

3: Primary flow 10-30% to UA - Micro Area Core

4: Primary flow within large Urban Cluster (10k - 49k)

4.1: Secondary flow 30-50% to UA - Micro High Commuting

5: Primary flow 30% or more to large UC

5.1: Secondary flow 30-50% to UA - Micro Low Commuting

6: Primary flow 10-30% to large UC - Small Town Core

7: Primary flow within small UC (2.5k - 9.9k)

7.1: Secondary flow 30-50% to UA

7.2: Secondary flow 30-50% to large UC - Small Town High Commuting

8: Primary flow 30% or more to small UC

8.1: Secondary flow 30-50% to UA

8.2: Secondary flow 30-50% to large UC - Small Town Low Commuting

9: Primary flow 10-30% to small UC - Rural Areas

10: Primary flow to tract outside a UA or UC

10.1: Secondary flow 30-50% to UA

10.2: Secondary flow 30-50% to large UC

10.3: Secondary flow 30-50% to small UC - 99: Zero population and no rural-urban identifier information

country

Country where provider is located.

specialty

Provider specialty code reported on the largest number of claims submitted.

par

Identifies whether the provider participates in Medicare and/or accepts assignment of Medicare allowed amounts. The value will be Y for any provider that had at least one claim identifying the provider as participating in Medicare or accepting assignment of Medicare allowed amounts within HCPCS code and place of service. A non-participating provider may elect to accept Medicare allowed amounts for some services and not accept Medicare allowed amounts for other services.

hcpcs_code

HCPCS code used to identify the specific medical service furnished by the provider. HCPCS codes include two levels. Level I codes are the Current Procedural Terminology (CPT) codes that are maintained by the American Medical Association and Level II codes are created by CMS to identify products, supplies and services not covered by the CPT codes (such as ambulance services).

drug

Identifies whether the HCPCS code for the specific service furnished by the provider is a HCPCS listed on the Medicare Part B Drug Average Sales Price (ASP) File.

pos

Identifies whether the place of service submitted on the claims is a facility ("F") or non-facility ("O"). Non-facility is generally an office setting; however other entities are included in non-facility.

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Details

The Provider and Service dataset provides information on services and procedures provided to Medicare (fee-for-service) Part B beneficiaries by physicians and other healthcare professionals.

The data is based on information gathered from CMS administrative claims data for Part B beneficiaries available from the CMS Chronic Conditions Data Warehouse.

The spending and utilization data in the Physician and Other Practitioners by Provider and Service Dataset are aggregated to the following:

the NPI for the performing provider,

the Healthcare Common Procedure Coding System (HCPCS) code, and

the place of service (either facility or non-facility).

There can be multiple records for a given NPI based on the number of distinct HCPCS codes that were billed and where the services were provided. Data have been aggregated based on the place of service because separate fee schedules apply depending on whether the place of service submitted on the claim is facility or non-facility.

Links

Medicare Physician & Other Practitioners: by Provider and Service API

Update Frequency: Annually

Examples

if (FALSE) { # interactive()
by_service(year = 2019, npi = 1003000126)
}