Search the Medicare Multiple Chronic Conditions API

Information on prevalence, use and spending by count of select chronic conditions among Original Medicare (or fee-for-service) beneficiaries.

Usage

cc_multiple(
  year,
  level = NULL,
  sublevel = NULL,
  fips = NULL,
  age_group = NULL,
  demographic = NULL,
  subdemo = NULL,
  tidy = TRUE
)

Format

In addition to the searchable columns:

mcc

To classify MCC for each Medicare beneficiary, the 21 chronic conditions are counted and grouped into four categories (0-1, 2-3, 4-5 and 6 or more).

prevalence

Prevalence estimates are calculated by taking the beneficiaries within the MCC category divided by the total number of beneficiaries in the fee-for-service population, expressed as a percentage.

tot_std_pymt_percap

Medicare standardized spending includes total Medicare payments for all covered services in Parts A and B and is presented per beneficiary (i.e. per capita). Standardized payments are presented to allow for comparisons across geographic areas in health care use among beneficiaries.

tot_pymt_percap

Medicare spending includes total Medicare payments for all covered services in Parts A and B and is presented per beneficiary (i.e. per capita).

hosp_readmsn_rate

Hospital readmissions are expressed as a percentage of all admissions. A 30-day readmission is defined as an admission to an acute care hospital for any cause within 30 days of discharge from an acute care hospital. Except when the patient died during the stay, each inpatient stay is classified as an index admission, a readmission, or both.

er_visits_per_1k

Emergency department visits are presented as the number of visits per 1,000 beneficiaries. ED visits include visits where the beneficiary was released from the outpatient setting and where the beneficiary was admitted to an inpatient setting.

Source

Centers for Medicare & Medicaid Services

Arguments

year

integer, YYYY, calendar year of Medicare enrollment. Run the helper function provider:::cc_multiple_years() to return a vector of currently available years.

level

Geographic level of data; options are National, State, and County

sublevel

The state and/or county where the Medicare beneficiary resides. The values include the 50 United States, District of Columbia, Puerto Rico or U.S. Virgin Islands. Data aggregated at the National level are identified by "National'.

fips

FIPS state and/or county code where the Medicare beneficiary resides. This column will be blank for data aggregated at the National level or for Puerto Rico and Virgin Islands.

age_group

Identifies the age level of the population that the data has been aggregated. A value of All indicates the data in the row represents all Fee-for-Service Medicare Beneficiaries. A value of <65 or 65+ indicates that the data is aggregated by the age of the Medicare Beneficiaries at the end of the calendar year.

demographic

Identifies the demographic level of the population that the data has been aggregated. A value of All indicates the data in the row represents all Fee-for-Service Medicare beneficiaries. A value of Sex indicates that the data has been aggregated by the Medicare beneficiary's sex. A value of Race indicates that the data has been aggregated by the Medicare beneficiary's race. A value of Dual Status indicates that the data has been aggregated by the Medicare beneficiary's dual eligibility status.

subdemo

For Sex, a beneficiary’s sex is classified as Male or Female and is identified using information from the CMS enrollment database. For Race, the race/ethnicity classifications are: Non-Hispanic White, Black or African American, Asian/Pacific Islander, Hispanic, and American Indian/Alaska Native. For Dual Status, beneficiaries can be classified as 'Medicare and Medicaid' or 'Medicare Only'. Beneficiaries enrolled in both Medicare and Medicaid are known as “dual eligibles.” Medicare beneficiaries are classified as dual eligibles if in any month in the given calendar year they were receiving full or partial Medicaid benefits.

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Details

The Multiple Chronic Conditions dataset provides information on the number of chronic conditions among Original Medicare beneficiaries. The dataset contains prevalence, use and spending organized by geography and the count of chronic conditions from the set of select 21 chronic conditions. The count of conditions is grouped into four categories (0-1, 2-3, 4-5 and 6 or more).

Links

Medicare Multiple Chronic Conditions

Note

Update Frequency: Annually

Examples

if (FALSE) {
cc_multiple(year = 2018, level  = "State", sublevel = "California")
cc_multiple(year = 2007, level = "National", demographic = "Race")
}