Search the Medicare Specific Chronic Conditions API

Information on prevalence, use and spending by count of select chronic conditions among Original Medicare (or fee-for-service) beneficiaries.

Usage

cc_specific(
  year = 2018,
  geo_lvl = NULL,
  geo_desc = NULL,
  fips = NULL,
  age_lvl = NULL,
  demo_lvl = NULL,
  demo_desc = NULL,
  condition = NULL,
  prevalence = NULL,
  stnd_pymt_pc = NULL,
  pymt_pc = NULL,
  readmit_rate = NULL,
  er_visits = NULL,
  clean_names = TRUE,
  lowercase = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

year

YYYY, calendar year of Medicare enrollment. 2007-2018 data is currently available.

geo_lvl

Geographic level of data; options are "National", "State", and "County"

geo_desc

The state and/or county where the Medicare beneficiary resides. The values include the 50 United States, District of Columbia, Puerto Rico or U.S. Virgin Islands. Data aggregated at the National level are identified by "National'.

fips

FIPS state and/or county code where the Medicare beneficiary resides. The Bene_Geo_Cd will be blank for data aggregated at the National level or for Puerto Rico and Virgin Islands.

age_lvl

Identifies the age level of the population that the data has been aggregated. A value of 'All' indicates the data in the row represents all Fee-for-Service Medicare Beneficiaries. A value of '<65' or '65+' indicates that the data is aggregated by the age of the Medicare Beneficiaries at the end of the calendar year.

demo_lvl

Identifies the demographic level of the population that the data has been aggregated. A value of 'All' indicates the data in the row is represents all Fee-for-Service Medicare beneficiaries. A value of 'Sex' indicates that the data has been aggregated by the Medicare beneficiary's sex. A value of 'Race' indicates that the data has been aggregated by the Medicare beneficiary's race. A value of 'Dual Status' indicates that the data has been aggregated by the Medicare beneficiary's dual eligibility status.

demo_desc

For Bene_Demo_Lvl='Sex', a beneficiary’s sex is classified as Male or Female and is identified using information from the CMS enrollment database. For Bene_Demo_Lvl='Race', the race/ethnicity classifications are: Non-Hispanic White, Black or African American, Asian/Pacific Islander, Hispanic, and American Indian/Alaska Native. All the chronic condition reports use the variable RTI_RACE_CD, which is available on the Master Beneficiary Files in the CCW. For Bene_Demo_Lvl='Dual Status',beneficiaries can be classified as 'Medicare & Medicaid' or 'Medicare Only'. Beneficiares enrolled in both Medicare and Medicaid are known as “dual eligibles.” Medicare beneficiaries are classified as dual eligibles if in any month in the given calendar year they were receiving full or partial Medicaid benefits.

condition

Identifies the chronic condition for which the prevalence and utilization is compiled. There are 21 chronic conditions identified using Medicare administrative claims. A Medicare beneficiary is considered to have a chronic condition if the CMS administrative data have a claim indicating that the beneficiary received a service or treatment for the specific condition. Beneficiaries may have more than one of the chronic conditions listed.

prevalence

Prevalence estimates are calculated by taking the beneficiaries within the MCC category divided by the total number of beneficiaries in the fee-for-service population, expressed as a percentage.

stnd_pymt_pc

Medicare standardized spending includes total Medicare payments for all covered services in Parts A and B and is presented per beneficiary (i.e. per capita). Standardized payments are presented to allow for comparisons across geographic areas in health care use among beneficiaries.

pymt_pc

Medicare spending includes total Medicare payments for all covered services in Parts A and B and is presented per beneficiary (i.e. per capita).

readmit_rate

Hospital readmissions are expressed as a percentage of all admissions. A 30-day readmission is defined as an admission to an acute care hospital for any cause within 30 days of discharge from an acute care hospital. Except when the patient died during the stay, each inpatient stay is classified as an index admission, a readmission, or both.

er_visits

Emergency department visits are presented as the number of visits per 1,000 beneficiaries. ED visits include visits where the beneficiary was released from the outpatient setting and where the beneficiary was admitted to an inpatient setting.

clean_names

Clean column names with janitor's clean_names() function; default is TRUE.

lowercase

Convert column names to lowercase; default is TRUE.

Value

A tibble containing the search results.

Details

The Select Chronic Conditions dataset provides information on 21 selected chronic conditions among Original Medicare beneficiaries. The dataset contains prevalence, use and spending organized by geography and distinct chronic conditions listed below.

• Alcohol Abuse Drug Abuse/ Substance Abuse • Alzheimer’s Disease and Related Dementia • Arthritis (Osteoarthritis and Rheumatoid) • Asthma • Atrial Fibrillation • Autism Spectrum Disorders • Cancer (Breast, Colorectal, Lung, and Prostate) • Chronic Kidney Disease • Chronic Obstructive Pulmonary Disease • Depression • Diabetes • Drug Abuse/ Substance Abuse • Heart Failure • Hepatitis (Chronic Viral B & C) • HIV/AIDS • Hyperlipidemia (High cholesterol) • Hypertension (High blood pressure) • Ischemic Heart Disease • Osteoporosis • Schizophrenia and Other Psychotic Disorders • Stroke

Links

Medicare Specific Chronic Conditions

Note

Update Frequency: Annually

Examples

if (FALSE) {
cc_specific(year     = 2018,
            geo_lvl  = "State",
            geo_desc = "California")

cc_specific(year     = 2007,
            geo_lvl  = "National",
            demo_lvl = "Race")
}