Clinical Laboratories

clia() allows you to search for information on clinical laboratories including demographics and the type of testing services the facility provides.

Clinical Laboratory Improvement Amendments (CLIA)

CMS regulates all laboratory testing (except research) performed on humans in the U.S. through the Clinical Laboratory Improvement Amendments (CLIA). In total, CLIA covers approximately 320,000 laboratory entities. The Division of Clinical Laboratory Improvement & Quality, within the Quality, Safety & Oversight Group, under the Center for Clinical Standards and Quality (CCSQ) has the responsibility for implementing the CLIA Program. Although all clinical laboratories must be properly certified to receive Medicare or Medicaid payments, CLIA has no direct Medicare or Medicaid program responsibilities.

Links:

Provider of Services File - Clinical Laboratories

Update Frequency: Quarterly

Usage

clia_labs(
  facility_name = NULL,
  clia = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  tidy = TRUE
)

Arguments

facility_name

Hospital’s CMS Certification Number (CCN)

clia

description

city

City of the hospital’s practice location address

state

State of the hospital’s practice location address

zip

State of the hospital’s practice location address

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Examples

if (FALSE) { # interactive()
clia_labs(city = "Valdosta")
clia_labs(clia = "11D0265516")
}