Chronic Conditions

conditions() allows the user access to data concerning chronic conditions among Original Medicare (or fee-for-service) beneficiaries.

Usage

conditions(
  year,
  condition = NULL,
  sublevel = NULL,
  type = c("multiple", "specific"),
  level = NULL,
  fips = NULL,
  age = NULL,
  demo = NULL,
  subdemo = NULL,
  mcc = NULL,
  tidy = TRUE,
  na.rm = TRUE
)

Arguments

year

< integer > // required Calendar year of Medicare enrollment, in YYYY format. Run cc_years() to return a vector of currently available years.

condition

< character > Chronic condition for which the prevalence and utilization is compiled (see above for list of conditions)

sublevel

< character > Beneficiary's state or county

type

"multiple" or "specific"

level

< character > Geographic level of aggregation: "national", "state", or "county"

fips

< character > Beneficiary's state or county FIPS code

age

< character > Age level of aggregation: "all", "<65", or "65+"

demo, subdemo

< character > Demographic/subdemographic level of aggregation: "all", "sex" ("male", "female"), "race" ("white", "black", "island", "hispanic", "native"), "dual" ("nondual", "dual").

mcc

< character > Number of chronic conditions: "0_1", "2_3", "4_5", "6_"

tidy

< boolean > // default: TRUE Tidy output

na.rm

< boolean > // default: TRUE Remove empty rows and columns

Value

A tibble with the following columns:

type = "specific"

The Specific Chronic Conditions dataset provides information on prevalence, use and spending organized by geography and 21 distinct chronic conditions among Original Medicare beneficiaries:

Alcohol Abuse

Alzheimer’s Disease/Dementia

Arthritis

Asthma

Atrial Fibrillation

Autism Spectrum Disorders

Cancer

Chronic Kidney Disease

COPD

Depression

Diabetes

Drug Abuse/Substance Abuse

Heart Failure

Hepatitis (Chronic Viral B & C)

HIV/AIDS

Hyperlipidemia

Hypertension

Ischemic Heart Disease

Osteoporosis

Schizophrenia and Other Psychotic Disorders

Stroke

type = "multiple"

The Multiple Chronic Conditions dataset provides information on prevalence, use and spending organized by geography and the count of chronic conditions from the set of select 21 chronic conditions. The count of conditions is grouped into four categories:

0-1

2-3

4-5

6 or More

Prevalence

Prevalence estimates are calculated by taking the beneficiaries within the MCC category divided by the total number of beneficiaries in the fee-for-service population, expressed as a percentage.

Hospital Readmission Rates

Hospital readmissions are expressed as a percentage of all admissions. A 30-day readmission is defined as an admission to an acute care hospital for any cause within 30 days of discharge from an acute care hospital. Except when the patient died during the stay, each inpatient stay is classified as an index admission, a readmission, or both.

Emergency Department Visits

Emergency department visits are presented as the number of visits per 1,000 beneficiaries. ED visits include visits where the beneficiary was released from the outpatient setting and where the beneficiary was admitted to an inpatient setting.

Payment Per Capita

Medicare spending includes total Medicare payments for all covered services in Parts A and B and is presented per beneficiary (i.e. per capita). Standardized payments are presented to allow for comparisons across geographic areas in health care use among beneficiaries.

Links

Medicare Multiple Chronic Conditions

Medicare Specific Chronic Conditions

Examples

if (FALSE) { # interactive()
chronic_conditions(year = 2018, type = "specific", level = "state", sublevel = "CA")
chronic_conditions(year = 2018, type = "multiple", level = "state", sublevel = "California")

chronic_conditions(year = 2007, type = "specific", level = "national", demo = "race")
chronic_conditions(year = 2007, type = "multiple", level = "national", demo = "race")
}