Search the CMS Doctors and Clinicians National Downloadable File API

Dataset of providers' facility affiliations

Usage

doctors_and_clinicians(
  npi = NULL,
  pac_id = NULL,
  enroll_id = NULL,
  first_name = NULL,
  middle_name = NULL,
  last_name = NULL,
  gender = NULL,
  school = NULL,
  grad_year = NULL,
  specialty = NULL,
  city = NULL,
  state = NULL,
  zipcode = NULL,
  ind_assn = NULL,
  group_assn = NULL,
  offset = 0,
  tidy = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

npi

Unique clinician ID assigned by NPPES

pac_id

Unique individual clinician ID assigned by PECOS

enroll_id

Unique ID for the clinician enrollment that is the source for the data in the observation

first_name

Individual clinician first name

middle_name

Individual clinician middle name

last_name

Individual clinician last name

gender

Individual clinician gender

school

Individual clinician’s medical school

grad_year

Individual clinician’s medical school graduation year

specialty

Primary medical specialty reported by the individual clinician in the selected enrollment

city

Group or individual's city

state

Group or individual's state

zipcode

Group or individual's ZIP code (9 digits when available)

ind_assn

Indicator for whether clinician accepts Medicare approved amount as payment in full. Y = Clinician accepts Medicare approved amount as payment in full. M = Clinician may accept Medicare Assignment.

group_assn

Indicator for whether group accepts Medicare approved amount as payment in full. Y = Clinician accepts Medicare approved amount as payment in full. M = Clinician may accept Medicare Assignment.

offset

offset; API pagination

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Details

The Doctors and Clinicians national downloadable file is organized such that each line is unique at the clinician/enrollment record/group/address level. Clinicians with multiple Medicare enrollment records and/or single enrollments linking to multiple practice locations are listed on multiple lines.

Links

Doctors and Clinicians National Downloadable File

Note

Update Frequency: Monthly

Examples

if (FALSE) { # interactive()
doctors_and_clinicians(npi = 1407263999)
doctors_and_clinicians(enroll_id = "I20081002000549")
doctors_and_clinicians(first_name = "John")
doctors_and_clinicians(school = "NEW YORK UNIVERSITY SCHOOL OF MEDICINE")
doctors_and_clinicians(grad_year = 2003)
}