Search the CMS Physician Facility Affiliations API

Dataset of providers' facility affiliations

Usage

facility_affiliations(
  npi = NULL,
  pac_id = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  facility_type = NULL,
  facility_ccn = NULL,
  parent_ccn = NULL,
  offset = 0,
  clean_names = TRUE,
  lowercase = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

npi

Unique clinician ID assigned by NPPES

pac_id

Unique individual clinician ID assigned by PECOS

last_name

Individual clinician last name

first_name

Individual clinician first name

middle_name

Individual clinician middle name

facility_type

Facilities can fall into the following type categories: Hospital, Long-term Care Hospital, Nursing Home, Inpatient Rehabilitation Facility, Home Health Agency, Hospice, and Dialysis Facility

facility_ccn

alphanumeric; Medicare CCN (CMS Certification Number) of facility type or unit within hospital where an individual clinician provides service. Effective 2007, the CCN replaced the terms Medicare Provider Number, Medicare Identification Number and OSCAR Number. The CCN is used to verify Medicare/Medicaid certification for survey and certification, assessment-related activities and communications.

parent_ccn

numeric; Medicare CCN of the primary hospital where an individual clinician provides service, should the clinician provide services in a unit within the hospital.

offset

offset; API pagination

clean_names

Clean column names with janitor's clean_names() function; default is TRUE.

lowercase

Convert column names to lowercase; default is TRUE.

Value

A tibble containing the search results.

Details

This file lists clinicians' facility affiliations, which have been expanded to include Long-term Care Hospitals (LTCHs), Skilled Nursing Facilities (SNFs), Inpatient Rehabilitation Facility (IRFs), Home Health Agencies, Hospices, and/or Dialysis Facilities, in addition to hospitals.

Links

CMS Physician Facility Affiliations

Certification Number (CCN) State Codes

Note

Update Frequency: Monthly

Examples

facility_affiliations(npi = 1003019563)
#> # A tibble: 4 × 10
#>   record_num…¹ npi   ind_p…² lst_nm frst_nm mid_nm suff  facil…³ facil…⁴ paren…⁵
#>   <chr>        <chr> <chr>   <chr>  <chr>   <chr>  <chr> <chr>   <chr>   <chr>  
#> 1 2541         1003… 468870… FRANK  JOHN    JOSEPH NA    Hospit… 060004  NA     
#> 2 2546         1003… 468870… FRANK  JOHN    JOSEPH NA    Hospit… 060009  NA     
#> 3 2551         1003… 468870… FRANK  JOHN    JOSEPH NA    Hospit… 060028  NA     
#> 4 2556         1003… 468870… FRANK  JOHN    JOSEPH NA    Hospit… 060116  NA     
#> # … with abbreviated variable names ¹​record_number, ²​ind_pac_id,
#> #   ³​facility_type, ⁴​facility_afl_ccn, ⁵​parent_ccn
facility_affiliations(facility_ccn = "060004")
#> # A tibble: 207 × 10
#>    record_nu…¹ npi   ind_p…² lst_nm frst_nm mid_nm suff  facil…³ facil…⁴ paren…⁵
#>    <chr>       <chr> <chr>   <chr>  <chr>   <chr>  <chr> <chr>   <chr>   <chr>  
#>  1 356         1003… 600295… HAMMAN DANIEL  RICHA… NA    Hospit… 060004  NA     
#>  2 2541        1003… 468870… FRANK  JOHN    JOSEPH NA    Hospit… 060004  NA     
#>  3 5431        1003… 589092… DONAH… ARTHUR  H      NA    Hospit… 060004  NA     
#>  4 13046       1003… 650718… HIGHA… JAMES   CHARL… NA    Hospit… 060004  NA     
#>  5 54036       1003… 125432… WEINER GARETH  R      NA    Hospit… 060004  NA     
#>  6 77556       1003… 549690… KLIMA  KATHRYN ANN    NA    Hospit… 060004  NA     
#>  7 85561       1013… 781009… HELZER AMITY   D      NA    Hospit… 060004  NA     
#>  8 90536       1013… 438589… PYLE   ASHLEY  L      NA    Hospit… 060004  NA     
#>  9 150361      1013… 155737… MUWAL… FIRAS   NA     NA    Hospit… 060004  NA     
#> 10 201526      1023… 286068… LUCCI  CHAD    MICHA… NA    Hospit… 060004  NA     
#> # … with 197 more rows, and abbreviated variable names ¹​record_number,
#> #   ²​ind_pac_id, ³​facility_type, ⁴​facility_afl_ccn, ⁵​parent_ccn
facility_affiliations(parent_ccn = 670055)
#> # A tibble: 11 × 10
#>    record_nu…¹ npi   ind_p…² lst_nm frst_nm mid_nm suff  facil…³ facil…⁴ paren…⁵
#>    <chr>       <chr> <chr>   <chr>  <chr>   <chr>  <chr> <chr>   <chr>   <chr>  
#>  1 676186      1083… 569883… FORNA… RAFAEL  JORGE  NA    Inpati… 67T055  670055 
#>  2 1132131     1144… 357765… CURRY  LYSA    LEE    NA    Inpati… 67T055  670055 
#>  3 3035336     1386… 327462… HAU    HORACIO G      NA    Inpati… 67T055  670055 
#>  4 3315471     1427… 145632… CHANEY DENNIS  M      NA    Inpati… 67T055  670055 
#>  5 4071956     1518… 862802… GUERRA ANTONIO F      NA    Inpati… 67T055  670055 
#>  6 4359616     1558… 892124… ALFON… JOHN    D      NA    Inpati… 67T055  670055 
#>  7 4777586     1609… 084037… WILCOX GEORGE  KIMBE… NA    Inpati… 67T055  670055 
#>  8 5571126     1710… 973933… VADDE… VIDYA   NA     NA    Inpati… 67T055  670055 
#>  9 5656051     1720… 701294… JANES  WILLIAM WARREN NA    Inpati… 67T055  670055 
#> 10 5847206     1740… 034547… WEIKLE GEOFF   R      NA    Inpati… 67T055  670055 
#> 11 7141806     1912… 509203… MANSO… SAAD    MD     NA    Inpati… 67T055  670055 
#> # … with abbreviated variable names ¹​record_number, ²​ind_pac_id,
#> #   ³​facility_type, ⁴​facility_afl_ccn, ⁵​parent_ccn
if (FALSE) {
facility_affiliations(first_name = "John")
facility_affiliations(facility_type = "Home Health Agency")
}