Providers with Missing Endpoints in NPPES

missing_information() allows you to search for providers with missing digital contact information in NPPES.

Missing Information = NPPES Endpoints

Digital contact information, also known as endpoints, provides a secure way for health care entities to send authenticated, encrypted health information to trusted recipients over the internet.

Health care organizations seeking to engage in electronic health information exchange need accurate information about the electronic addresses (e.g., Direct address, FHIR server URL, query endpoint, or other digital contact information) of potential exchange partners to facilitate this information exchange.

Usage

missing_information(npi = NULL, name = NULL, tidy = TRUE)

Arguments

npi

The provider’s National Provider Identifier

name

Provider's full name, in the form "last, first"

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Links

CMS Public Reporting of Missing Digital Contact Information API

Endpoints Information

Methodology & Policy

Update Frequency

Quarterly

See also

Examples

# A provider that appears in the search results of the Missing Information
# API has no Endpoints entered into the NPPES NPI Registry and vice versa.

## Appears
missing_information(name = "Clouse, John")
#> # A tibble: 1 × 3
#>   npi        last_name first_name
#>   <chr>      <chr>     <chr>     
#> 1 1144224569 Clouse    John      

## No Endpoints in NPPES
nppes_npi(npi = 1144224569,
          tidy = FALSE) |>
          dplyr::select(endpoints)
#> # A tibble: 1 × 1
#>   endpoints
#>   <lgl>    
#> 1 NA       

## Does Not Appear
missing_information(npi = 1003000423)
#> ✖ No results for npi: 1003000423

## Has Endpoints in NPPES
nppes_npi(npi = 1003000423, tidy = FALSE) |>
dplyr::select(endpoints) |>
tidyr::unnest(cols = c(endpoints)) |>
janitor::clean_names() |>
dplyr::select(dplyr::contains("endpoint"))
#> # A tibble: 1 × 4
#>   endpoint_type endpoint_type_description endpoint          endpoint_description
#>   <chr>         <chr>                     <chr>             <chr>               
#> 1 DIRECT        Direct Messaging Address  Jennifer.velotta… eclinicaldirectplus