Search the NPPES National Provider Identifier Registry API

nppes_npi() allows you to search the NPPES NPI Registry's public API by many of the parameters defined in the API's documentation.

Usage

nppes_npi(
  npi = NULL,
  entype = NULL,
  first_name = NULL,
  last_name = NULL,
  org_name = NULL,
  purpose_name = NULL,
  taxonomy_desc = NULL,
  city = NULL,
  state = NULL,
  zipcode = NULL,
  country = NULL,
  limit = 1200,
  skip = NULL,
  tidy = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

npi

10-digit National Provider Identifier (NPI).

entype

Choices are either I for an Individual provider (NPI-1 or Type 1) or O for an Organizational provider (NPI-2 or Type 2.). When not specified, both Type 1 and Type 2 NPIs will be returned. Entype cannot be the only criteria entered.

first_name

Individual provider's (NPI-1) first name. Trailing wildcard entries are permitted requiring at least two characters to be entered (e.g. jo* ).

last_name

Individual provider's (NPI-1) last name. Trailing wildcard entries are permitted requiring at least two characters to be entered (e.g. jo* ).

org_name

Healthcare organization's name (NPI-2). Trailing wildcard entries are permitted requiring at least two characters to be entered. All types of Organization Names (LBN, DBA, Former LBN, Other Name) associated with an NPI are examined for matching contents, therefore, the results might contain an organization name different from the one entered in the Organization Name criterion.

purpose_name

Refers to whether the name information entered pertains to an Authorized Official's name or a Provider's name. When not specified, the results will search against a provider's first and last name. AO will only search against Authorized Official names. While PROVIDER will only search against Provider name. Valid values are: AO and Provider.

taxonomy_desc

Search for providers by their taxonomy by entering the taxonomy description.

city

City associated with the provider's address. To search for a Military Address, enter either APO or FPO.

state

State abbreviation associated with the provider's address. If this field is used, at least one other field, besides the entype and country, must be populated. Valid values for state abbreviations: State Abbreviations.

zipcode

The Postal Code associated with the provider's address identified in Address Purpose. If you enter a 5 digit postal code, it will match any appropriate 9 digit (zip+4) codes in the data. Trailing wildcard entries are permitted requiring at least two characters to be entered (e.g., 21*).

country

Country abbreviation associated with the provider's address. This field can be used as the only input criterion, as long as the value selected is not US (United States). Valid values for country abbreviations: Country Abbreviations.

limit

Maximum number of results to return; default is 1200.

skip

Number of results to skip after searching the previous number; set in limit.

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Details

The NPPES NPI Registry Public Search is a free directory of all active National Provider Identifier (NPI) records. Healthcare providers acquire their unique 10-digit NPIs to identify themselves in a standard way throughout their industry. After CMS supplies an NPI, they publish the parts of the NPI record that have public relevance, including the provider’s name, taxonomy and practice address. It enables you to search for providers in the NPPES (National Plan and Provider Enumeration System.) All information produced by the NPI Registry is provided in accordance with the NPPES Data Dissemination Notice. There is no charge to use the NPI Registry.

Links

NPPES NPI Registry API Documentation

NPPES NPI Registry API Demo

Note

Update Frequency: Weekly

Examples

if (FALSE) { # interactive()
nppes_full(npi = 1528060837)
}