Search CMS' Open Payments Program API

open_payments() allows you to search CMS' Open Payments Program API.

Usage

open_payments(
  year,
  npi = NULL,
  profile_id = NULL,
  covered_type = NULL,
  first_name = NULL,
  last_name = NULL,
  city = NULL,
  state = NULL,
  zipcode = NULL,
  teaching_hospital = NULL,
  payer_name = NULL,
  payer_id = NULL,
  payment_form = NULL,
  payment_nature = NULL,
  offset = 0,
  tidy = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

year

Program reporting year. Run the helper function provider:::open_payments_years() to return a vector of currently available years.

npi

Covered recipient's National Provider Identifier (NPI).

profile_id

Covered recipient's unique Open Payments ID

covered_type

Type of covered recipient, e.g.,

Covered Recipient Physician

Covered Recipient Non-Physician Practitioner

Covered Recipient Teaching Hospital

first_name

Covered recipient's first name

last_name

Covered recipient's last name

city

City

state

State, abbreviation

zipcode

Zip code

teaching_hospital

Name of teaching hospital, e.g. Vanderbilt University Medical Center

payer_name

Paying entity's name. Examples:

Pharmacosmos Therapeutics Inc.

Getinge USA Sales, LLC

Agiliti Health, Inc.

OrthoScan, Inc.

payer_id

Paying entity's unique Open Payments ID

payment_form

Form of payment, examples:

Stock option

Cash or cash equivalent

In-kind items and services

payment_nature

Nature of payment or transfer of value, examples:

Royalty or License

Charitable Contribution

Current or prospective ownership or investment interest

Food and Beverage

offset

offset; API pagination

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Details

The Open Payments program is a national disclosure program that promotes a more transparent and accountable health care system. Open Payments houses a publicly accessible database of payments that reporting entities, including drug and medical device companies, make to covered recipients like physicians. Please note that CMS does not comment on what relationships may be beneficial or potential conflicts of interest. CMS publishes the data attested to by reporting entities. The data is open to individual interpretation.

Open Payments is a national transparency program that collects and publishes information about financial relationships between drug and medical device companies (referred to as "reporting entities") and certain health care providers (referred to as "covered recipients"). These relationships may involve payments to providers for things including but not limited to research, meals, travel, gifts or speaking fees.

The purpose of the program is to provide the public with a more transparent health care system. All information available on the Open Payments database is open to personal interpretation and if there are questions about what the data means, patients and their advocates should speak directly to the health care provider for a better understanding.

Links

What is the Open Payments Program?

OpenPaymentsData.cms.gov

2021 General Payment Data

2020 General Payment Data

2019 General Payment Data

2018 General Payment Data

2017 General Payment Data

2016 General Payment Data

2015 General Payment Data

Note

Update Frequency: Yearly

Examples

if (FALSE) { # interactive()
open_payments(npi = 1043218118, year = 2021)
open_payments(payment_nature = "Royalty or License")
open_payments(payment_form = "Stock option")
open_payments(payer_name = "Adaptive Biotechnologies Corporation")
open_payments(teaching_hospital = "Nyu Langone Hospitals")
}