Order and Referral Eligibility

order_refer() returns a provider's eligibility to order and refer within Medicare to:

Usage

order_refer(
  npi = NULL,
  first = NULL,
  last = NULL,
  partb = NULL,
  dme = NULL,
  hha = NULL,
  pmd = NULL,
  hos = NULL,
  tidy = TRUE,
  pivot = TRUE,
  ...
)

Arguments

npi

<chr> Unique 10-digit National Provider Identifier number issued by CMS to US healthcare providers through NPPES.

first, last

<chr> Individual provider's first/last name

partb, dme, hha, pmd, hos

<lgl> Whether a provider is eligible to order and refer to:

partb: Medicare Part B

dme: Durable Medical Equipment

hha: Home Health Agency

pmd: Power Mobility Devices

hos: Hospice

tidy

<lgl> Tidy output; default is TRUE

pivot

<lgl> Pivot output; default is TRUE

...

Empty dots

Value

A tibble with the columns:

Field	Description

npi	National Provider Identifier

first	Order and Referring Provider's First Name

last	Order and Referring Provider's Last Name

eligible	Services An Eligible Provider Can Order/Refer To

Details

Part B: Clinical Laboratory Services, Imaging Services

DME: Durable Medical Equipment, Prosthetics, Orthotics, & Supplies (DMEPOS)

Part A: Home Health Services

To be eligible, a provider must:

have an Individual NPI

be enrolled in Medicare in either an Approved or Opt-Out status

be of an Eligible Specialty type

Ordering Providers can order non-physician services for patients.

Referring (or Certifying) Providers can request items or services that Medicare may reimburse on behalf of its beneficiaries.

Opt-Out Providers: Providers who have opted out of Medicare may still order and refer. They can also enroll solely to order and refer.

References

links:

Medicare Order and Referring API

CMS.gov: Ordering & Certifying

Order and Referring Methodology

Update Frequency: Twice Weekly

Examples

order_refer(npi = 1003026055)
#> # A tibble: 5 × 4
#>   npi        first   last   eligible                 
#>   <chr>      <chr>   <chr>  <fct>                    
#> 1 1003026055 RADHIKA PHADKE Medicare Part B          
#> 2 1003026055 RADHIKA PHADKE Home Health Agency       
#> 3 1003026055 RADHIKA PHADKE Durable Medical Equipment
#> 4 1003026055 RADHIKA PHADKE Power Mobility Devices   
#> 5 1003026055 RADHIKA PHADKE Hospice                  

# Filter for certain privileges
order_refer(first = "Jennifer",
            last  = "Smith",
            partb = TRUE,
            hos   = FALSE,
            hha   = FALSE,
            pmd   = FALSE)
#> # A tibble: 40 × 4
#>    npi        first    last  eligible                 
#>    <chr>      <chr>    <chr> <fct>                    
#>  1 1992162556 JENNIFER SMITH Medicare Part B          
#>  2 1992162556 JENNIFER SMITH Durable Medical Equipment
#>  3 1811518566 JENNIFER SMITH Medicare Part B          
#>  4 1811518566 JENNIFER SMITH Durable Medical Equipment
#>  5 1205226172 JENNIFER SMITH Medicare Part B          
#>  6 1205226172 JENNIFER SMITH Durable Medical Equipment
#>  7 1922644848 JENNIFER SMITH Medicare Part B          
#>  8 1922644848 JENNIFER SMITH Durable Medical Equipment
#>  9 1447039060 JENNIFER SMITH Medicare Part B          
#> 10 1447039060 JENNIFER SMITH Durable Medical Equipment
#> # ℹ 30 more rows