Search the Medicare Pending Initial Logging and Tracking API

A list of enrollment applications pending CMS contractor review for physicians & non-physicians.

Usage

pending_applications(
  npi = NULL,
  last_name = NULL,
  first_name = NULL,
  type = c("physician", "non-physician"),
  tidy = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

npi

National Provider Identifier (NPI) number

last_name

Last name of provider

first_name

First name of provider

type

physician or non-physician

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Details

The Pending Initial Logging and Tracking (L & T) dataset provides a list of pending applications that have not been processed by CMS contractors.

Links

Medicare Pending Initial Logging and Tracking Physicians API

Medicare Pending Initial Logging and Tracking Non-Physicians API

Note

Update Frequency: Weekly

Examples

pending_applications(last_name = "Smith", type = "non-physician")
#> # A tibble: 23 × 4
#>    npi        last_name first_name type         
#>    <chr>      <chr>     <chr>      <chr>        
#>  1 1518672245 SMITH     ALEXANDRA  NON-PHYSICIAN
#>  2 1700588621 SMITH     AMANDA     NON-PHYSICIAN
#>  3 1073234282 SMITH     ANGELA     NON-PHYSICIAN
#>  4 1790408185 SMITH     ARIELLE    NON-PHYSICIAN
#>  5 1801992912 SMITH     BRIDGET    NON-PHYSICIAN
#>  6 1558504688 SMITH     BRITTANY   NON-PHYSICIAN
#>  7 1639635030 SMITH     COURTNEY   NON-PHYSICIAN
#>  8 1043925035 SMITH     FULGENTIA  NON-PHYSICIAN
#>  9 1902527641 SMITH     HEATHER    NON-PHYSICIAN
#> 10 1720545619 SMITH     HILLARY    NON-PHYSICIAN
#> # ℹ 13 more rows
pending_applications(first_name = "John", type = "physician")
#> # A tibble: 35 × 4
#>    npi        last_name  first_name type     
#>    <chr>      <chr>      <chr>      <chr>    
#>  1 1881791739 ADAMS      JOHN       PHYSICIAN
#>  2 1841280963 BIGBEE     JOHN       PHYSICIAN
#>  3 1619996378 BODDEN     JOHN       PHYSICIAN
#>  4 1588744569 BRUNO      JOHN       PHYSICIAN
#>  5 1861142556 BURKE      JOHN       PHYSICIAN
#>  6 1306817531 COMBS      JOHN       PHYSICIAN
#>  7 1730349580 ECHEVARRIA JOHN       PHYSICIAN
#>  8 1659074151 EUN        JOHN       PHYSICIAN
#>  9 1376571554 FLYNN      JOHN       PHYSICIAN
#> 10 1689774804 FREEMAN    JOHN       PHYSICIAN
#> # ℹ 25 more rows