Yearly Utilization Data on Medicare Part B Providers

physician_by_provider() allows you to access data such as services and procedures performed; charges submitted and payment received; and beneficiary demographic and health characteristics for providers treating Original Medicare (fee-for-service) Part B beneficiaries, aggregated by year.

Usage

physician_by_provider(
  year,
  npi = NULL,
  first_name = NULL,
  last_name = NULL,
  organization_name = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zipcode = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par = NULL,
  tidy = TRUE
)

physician_by_provider_years()

Arguments

year

integer (required); Year in YYYY format. Run helper function physician_by_provider_years() to return a vector of the years currently available.

npi

National Provider Identifier for the rendering provider on the claim.

first_name

Individual provider's first name.

last_name

Individual provider's last name.

organization_name

Organization name.

credential

Individual provider's credentials.

gender

Individual provider's gender.

entype

Provider entity type.

city

City where provider is located.

state

State where provider is located.

zipcode

Provider’s zip code.

fips

Provider's state FIPS code.

ruca

Rural-Urban Commuting Area Code (RUCA); a Census tract-based classification scheme that utilizes the standard Bureau of Census Urbanized Area and Urban Cluster definitions in combination with work commuting information to characterize all of the nation's Census tracts regarding their rural and urban status and relationships. The Referring Provider ZIP code was cross walked to the United States Department of Agriculture (USDA) 2010 Rural-Urban Commuting Area Codes.

country

Country where provider is located.

specialty

Provider specialty code reported on the largest number of claims submitted.

par

Identifies whether the provider participates in Medicare and/or accepts assignment of Medicare allowed amounts. The value will be Y for any provider that had at least one claim identifying the provider as participating in Medicare or accepting assignment of Medicare allowed amounts within HCPCS code and place of service. A non-participating provider may elect to accept Medicare allowed amounts for some services and not accept Medicare allowed amounts for other services.

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

integer vector of years available

Links

Medicare Physician & Other Practitioners: by Provider API

Update Frequency

*Annually

See also

Examples

if (FALSE) { # interactive()
physician_by_provider(year = 2020,
                      npi = 1003000423)

# Use the years helper function to retrieve results for every year:
physician_by_provider_years() |>
map(\(x) physician_by_provider(year = x, npi = 1043477615)) |>
list_rbind()
}
physician_by_provider_years()
#> [1] 2013 2014 2015 2016 2017 2018 2019 2020 2021