Search the Medicare Physician & Other Practitioners API by Provider and Service

Information on services and procedures provided to Original Medicare (fee-for-service) Part B (Medical Insurance) beneficiaries by physicians and other healthcare professionals; aggregated by provider and service.

Usage

physician_by_service(
  npi = NULL,
  last_org = NULL,
  first = NULL,
  cred = NULL,
  gender = NULL,
  type = NULL,
  city = NULL,
  state = NULL,
  fips = NULL,
  zip = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par_ind = NULL,
  hcpcs_code = NULL,
  hcpcs_drug = NULL,
  pos = NULL,
  year = 2020,
  clean_names = TRUE,
  lowercase = TRUE,
  nest = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

npi

National Provider Identifier (NPI) for the rendering provider on the claim. The provider NPI is the numeric identifier registered in NPPES.

last_org

Last name/Organization name of the provider. When the provider is registered in NPPES as an individual (entity type code=’I’), this is the provider’s last name. When the provider is registered as an organization (entity type code = ‘O’), this is the organization name.

first

An individual provider's (entity type code=’I’) first name. An organization's (entity type code = ‘O’) will be blank.

cred

An individual provider's (entity type code=’I’) credentials. An organization's will be blank.

gender

An individual provider's gender. An organization's will be blank.

type

Type of entity reported in NPPES. An entity code of ‘I’ identifies providers registered as individuals while an entity type code of ‘O’ identifies providers registered as organizations.

city

The city where the provider is located, as reported in NPPES.

state

The state where the provider is located, as reported in NPPES.

fips

FIPS code for the rendering provider's state.

zip

The provider’s zip code, as reported in NPPES.

ruca

Rural-Urban Commuting Area Code (RUCA); a Census tract-based classification scheme that utilizes the standard Bureau of Census Urbanized Area and Urban Cluster definitions in combination with work commuting information to characterize all of the nation's Census tracts regarding their rural and urban status and relationships. The Referring Provider ZIP code was cross walked to the United States Department of Agriculture (USDA) 2010 Rural-Urban Commuting Area Codes.

country

The country where the provider is located, as reported in NPPES.

specialty

Derived from the provider specialty code reported on the claim. For providers that reported more than one specialty code on their claims, this is the specialty code associated with the largest number of services.

par_ind

Identifies whether the provider participates in Medicare and/or accepts assignment of Medicare allowed amounts. The value will be ‘Y’ for any provider that had at least one claim identifying the provider as participating in Medicare or accepting assignment of Medicare allowed amounts within HCPCS code and place of service. A non-participating provider may elect to accept Medicare allowed amounts for some services and not accept Medicare allowed amounts for other services.

hcpcs_code

HCPCS code used to identify the specific medical service furnished by the provider. HCPCS codes include two levels. Level I codes are the Current Procedural Terminology (CPT) codes that are maintained by the American Medical Association and Level II codes are created by CMS to identify products, supplies and services not covered by the CPT codes (such as ambulance services).

hcpcs_drug

Identifies whether the HCPCS code for the specific service furnished by the provider is a HCPCS listed on the Medicare Part B Drug Average Sales Price (ASP) File. Please visit the ASP drug pricing page for additional information.

pos

Identifies whether the place of service submitted on the claims is a facility (value of ‘F’) or non-facility (value of ‘O’). Non-facility is generally an office setting; however other entities are included in non-facility.

year

Year in YYYY format, between 2013-2020; default is 2020

clean_names

Clean column names with janitor's clean_names() function; default is TRUE.

lowercase

Convert column names to lowercase; default is TRUE.

nest

Nest related columns together; default is TRUE.

Value

A tibble containing the search results.

Details

The Provider and Service dataset provides information on services and procedures provided to Medicare (fee-for-service) Part B beneficiaries by physicians and other healthcare professionals.

The data is based on information gathered from CMS administrative claims data for Part B beneficiaries available from the CMS Chronic Conditions Data Warehouse.

The spending and utilization data in the Physician and Other Practitioners by Provider and Service Dataset are aggregated to the following:

the NPI for the performing provider,

the Healthcare Common Procedure Coding System (HCPCS) code, and

the place of service (either facility or non-facility).

There can be multiple records for a given NPI based on the number of distinct HCPCS codes that were billed and where the services were provided. Data have been aggregated based on the place of service because separate fee schedules apply depending on whether the place of service submitted on the claim is facility or non-facility.

Links

Medicare Physician & Other Practitioners: by Provider and Service API

Note

Update Frequency: Annually

Examples

if (FALSE) {
# Search by NPI
physician_by_service(npi = 1003000126)

# Search by Last Name for 2019
physician_by_service(last_org = "Enkeshafi", year = 2019)

# Multiple NPIs
npis <- c(1003026055,
          1316405939,
          1720392988,
          1518184605,
          1922056829,
          1083879860)

npis |> purrr::map_dfr(physician_by_service)

# Retrieve All Provider Data, 2013-2020
purrr::map_dfr(as.character(2013:2020),
~physician_by_service(npi = 1003000126, year = .x))
}