Search the NPPES NPI Public Registry

The NPPES NPI Registry Public Search is a free directory of all active National Provider Identifier (NPI) records. Healthcare providers acquire their unique 10-digit NPIs to identify themselves in a standard way throughout their industry. After CMS supplies an NPI, they publish the parts of the NPI record that have public relevance, including the provider’s name, taxonomy and practice address. It enables you to search for providers in the NPPES (National Plan and Provider Enumeration System.) All information produced by the NPI Registry is provided in accordance with the NPPES Data Dissemination Notice. There is no charge to use the NPI Registry.

Usage

prov_city_nppes(city, state, limit = 200, skip = NULL)

Arguments

city

city in which provider practices

state

abbreviation of state in which provider practices

limit

maximum number of results expected back

skip

number of results to skip after searching the previous, say, 200 examples

Value

A tibble containing the NPI(s) searched for, the date-time the search was performed, and a list-column of the results.

Examples

prov_city_nppes(city = "Atlanta", state = "GA", limit = "1")
#> Warning: Event: on_file_type_mismatch
#> Warning: Event: on_suspect_content
#> # A tibble: 1 × 3
#>   datetime            outcome data_lists   
#>   <dttm>              <chr>   <list>       
#> 1 2022-09-16 10:09:35 results <df [1 × 11]>