Search the Medicare Physician & Other Practitioners - by Provider and Service API

The Medicare Physician & Other Practitioners by Provider and Service dataset provides information on use, payments, and submitted charges organized by National Provider Identifier (NPI), Healthcare Common Procedure Coding System (HCPCS) code, and place of service.

Usage

prov_mcr_phys_pract(npi, year = "2020")

Source

https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider-and-service

Arguments

npi

10-digit National Provider Identifier (NPI)

year

year between 2013-2020, in YYYY format

Value

A tibble containing the NPI(s) searched for, the date-time the search was performed, and a list-column of the results.

Details

Information on services and procedures provided to Original Medicare (fee-for-service) Part B (Medical Insurance) beneficiaries by physicians and other healthcare professionals; aggregated by provider and service.

Examples

prov_mcr_phys_pract(npi = 1003000126, year = "2020")
#>   year Rndrng_NPI Rndrng_Prvdr_Last_Org_Name Rndrng_Prvdr_First_Name
#> 1 2020 1003000126                  Enkeshafi                 Ardalan
#> 2 2020 1003000126                  Enkeshafi                 Ardalan
#> 3 2020 1003000126                  Enkeshafi                 Ardalan
#> 4 2020 1003000126                  Enkeshafi                 Ardalan
#> 5 2020 1003000126                  Enkeshafi                 Ardalan
#> 6 2020 1003000126                  Enkeshafi                 Ardalan
#> 7 2020 1003000126                  Enkeshafi                 Ardalan
#> 8 2020 1003000126                  Enkeshafi                 Ardalan
#> 9 2020 1003000126                  Enkeshafi                 Ardalan
#>   Rndrng_Prvdr_MI Rndrng_Prvdr_Crdntls Rndrng_Prvdr_Gndr Rndrng_Prvdr_Ent_Cd
#> 1                                 M.D.                 M                   I
#> 2                                 M.D.                 M                   I
#> 3                                 M.D.                 M                   I
#> 4                                 M.D.                 M                   I
#> 5                                 M.D.                 M                   I
#> 6                                 M.D.                 M                   I
#> 7                                 M.D.                 M                   I
#> 8                                 M.D.                 M                   I
#> 9                                 M.D.                 M                   I
#>            Rndrng_Prvdr_St1 Rndrng_Prvdr_St2 Rndrng_Prvdr_City
#> 1 6410 Rockledge Dr Ste 304                           Bethesda
#> 2 6410 Rockledge Dr Ste 304                           Bethesda
#> 3 6410 Rockledge Dr Ste 304                           Bethesda
#> 4 6410 Rockledge Dr Ste 304                           Bethesda
#> 5 6410 Rockledge Dr Ste 304                           Bethesda
#> 6 6410 Rockledge Dr Ste 304                           Bethesda
#> 7 6410 Rockledge Dr Ste 304                           Bethesda
#> 8 6410 Rockledge Dr Ste 304                           Bethesda
#> 9 6410 Rockledge Dr Ste 304                           Bethesda
#>   Rndrng_Prvdr_State_Abrvtn Rndrng_Prvdr_State_FIPS Rndrng_Prvdr_Zip5
#> 1                        MD                      24             20817
#> 2                        MD                      24             20817
#> 3                        MD                      24             20817
#> 4                        MD                      24             20817
#> 5                        MD                      24             20817
#> 6                        MD                      24             20817
#> 7                        MD                      24             20817
#> 8                        MD                      24             20817
#> 9                        MD                      24             20817
#>   Rndrng_Prvdr_RUCA
#> 1                 1
#> 2                 1
#> 3                 1
#> 4                 1
#> 5                 1
#> 6                 1
#> 7                 1
#> 8                 1
#> 9                 1
#>                                                                Rndrng_Prvdr_RUCA_Desc
#> 1 Metropolitan area core: primary flow within an urbanized area of 50,000 and greater
#> 2 Metropolitan area core: primary flow within an urbanized area of 50,000 and greater
#> 3 Metropolitan area core: primary flow within an urbanized area of 50,000 and greater
#> 4 Metropolitan area core: primary flow within an urbanized area of 50,000 and greater
#> 5 Metropolitan area core: primary flow within an urbanized area of 50,000 and greater
#> 6 Metropolitan area core: primary flow within an urbanized area of 50,000 and greater
#> 7 Metropolitan area core: primary flow within an urbanized area of 50,000 and greater
#> 8 Metropolitan area core: primary flow within an urbanized area of 50,000 and greater
#> 9 Metropolitan area core: primary flow within an urbanized area of 50,000 and greater
#>   Rndrng_Prvdr_Cntry Rndrng_Prvdr_Type Rndrng_Prvdr_Mdcr_Prtcptg_Ind HCPCS_Cd
#> 1                 US Internal Medicine                             Y    99217
#> 2                 US Internal Medicine                             Y    99218
#> 3                 US Internal Medicine                             Y    99220
#> 4                 US Internal Medicine                             Y    99221
#> 5                 US Internal Medicine                             Y    99223
#> 6                 US Internal Medicine                             Y    99232
#> 7                 US Internal Medicine                             Y    99233
#> 8                 US Internal Medicine                             Y    99238
#> 9                 US Internal Medicine                             Y    99239
#>                                                         HCPCS_Desc
#> 1                    Hospital observation care on day of discharge
#> 2                  Hospital observation care, typically 30 minutes
#> 3                  Hospital observation care, typically 70 minutes
#> 4    Initial hospital inpatient care, typically 30 minutes per day
#> 5    Initial hospital inpatient care, typically 70 minutes per day
#> 6 Subsequent hospital inpatient care, typically 25 minutes per day
#> 7 Subsequent hospital inpatient care, typically 35 minutes per day
#> 8            Hospital discharge day management, 30 minutes or less
#> 9          Hospital discharge day management, more than 30 minutes
#>   HCPCS_Drug_Ind Place_Of_Srvc Tot_Benes Tot_Srvcs Tot_Bene_Day_Srvcs
#> 1              N             F        23        23                 23
#> 2              N             F        16        16                 16
#> 3              N             F        16        16                 16
#> 4              N             F        12        12                 12
#> 5              N             F        52        52                 52
#> 6              N             F        57       134                134
#> 7              N             F       154       326                326
#> 8              N             F        13        13                 13
#> 9              N             F       145       146                146
#>   Avg_Sbmtd_Chrg Avg_Mdcr_Alowd_Amt Avg_Mdcr_Pymt_Amt Avg_Mdcr_Stdzd_Amt
#> 1         406.17              76.80             61.40              58.00
#> 2         811.81             108.25             85.08              78.32
#> 3        1150.94             191.96            152.38             146.61
#> 4         681.67             110.05             82.24              76.40
#> 5        1108.38             209.42            170.83             164.66
#> 6         251.33              73.45             57.96              56.98
#> 7         513.21             108.85             88.95              84.91
#> 8         268.46              74.16             60.53              59.41
#> 9         491.87             111.26             90.92              87.21