Search the Medicare Fee-For-Service Public Provider Enrollment API

Information on a point in time snapshot of enrollment level data for providers actively enrolled in Medicare.

Usage

provider_enrollment(
  npi = NULL,
  pecos_id = NULL,
  enroll_id = NULL,
  prov_type_code = NULL,
  prov_type_desc = NULL,
  state = NULL,
  first_name = NULL,
  middle_name = NULL,
  last_name = NULL,
  org_name = NULL,
  gender = NULL,
  clean_names = TRUE,
  lowercase = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

npi

An NPI is a 10-digit unique numeric identifier that all providers must obtain before enrolling in Medicare. It is assigned to health care providers upon application through the National Plan and Provider Enumeration System (NPPES).

pecos_id

Provider associate level variable (PAC ID) from PECOS database used to link across tables. A PAC ID is a 10-digit unique numeric identifier that is assigned to each individual or organization in PECOS. All entity-level information (e.g., tax identification numbers and organizational names) is linked through the PAC ID. A PAC ID may be associated with multiple Enrollment IDs if the individual or organization enrolled multiple times under different circumstances.

enroll_id

Provider enrollment ID from PECOS database used to link across tables. An Enrollment ID is a 15-digit unique alphanumeric identifier that is assigned to each new provider enrollment application. All enrollment-level information (e.g., enrollment type, enrollment state, provider specialty and reassignment of benefits) is linked through the Enrollment ID.

prov_type_code

Provider enrollment application and enrollment specialty type. This field shows the provider’s primary specialty code. For practitioners and DME suppliers, please see the Secondary Specialty file for a list of secondary specialties (when applicable). Only about 20% of practitioners and DME suppliers have at least one secondary specialty.

prov_type_desc

Provider enrollment application and enrollment specialty type description

state

Provider enrollment state, abbreviated location. Providers enroll at the state level, so one PAC ID may be associated with multiple ENRLMT_IDs and multiple STATE_CD values.

first_name

Individual provider first name

middle_name

Individual provider middle name

last_name

Individual provider last name

org_name

Organizational provider name

gender

Individual provider gender; F(female), M(male), 9(unknown)

clean_names

Clean column names with janitor's clean_names() function; default is TRUE.

lowercase

Convert column names to lowercase; default is TRUE.

Value

A tibble containing the search results.

Details

The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are actively approved to bill Medicare or have completed the 855O at the time the data was pulled from the Provider Enrollment and Chain Ownership System (PECOS). The release of this provider enrollment data is not related to other provider information releases such as Physician Compare or Data Transparency.

Links

Medicare Fee-For-Service Public Provider Enrollment API

Note

Update Frequency: Quarterly

Examples

provider_enrollment(npi = 1003026055)
#> # A tibble: 1 × 11
#>   npi    pecos…¹ enrlm…² provi…³ provi…⁴ state…⁵ first…⁶ mdl_n…⁷ last_…⁸ org_n…⁹
#>   <chr>  <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 10030… 448682… I20170… 14-46   PRACTI… FL      RADHIKA PUSHKAR PHADKE  ""     
#> # … with 1 more variable: gndr_sw <chr>, and abbreviated variable names
#> #   ¹​pecos_asct_cntl_id, ²​enrlmt_id, ³​provider_type_cd, ⁴​provider_type_desc,
#> #   ⁵​state_cd, ⁶​first_name, ⁷​mdl_name, ⁸​last_name, ⁹​org_name
provider_enrollment(first_name = "DEBRA", middle_name = "L", last_name = "FROMER")
#> # A tibble: 1 × 11
#>   npi    pecos…¹ enrlm…² provi…³ provi…⁴ state…⁵ first…⁶ mdl_n…⁷ last_…⁸ org_n…⁹
#>   <chr>  <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 18018… 135525… I20031… 14-34   PRACTI… NJ      DEBRA   L       FROMER  ""     
#> # … with 1 more variable: gndr_sw <chr>, and abbreviated variable names
#> #   ¹​pecos_asct_cntl_id, ²​enrlmt_id, ³​provider_type_cd, ⁴​provider_type_desc,
#> #   ⁵​state_cd, ⁶​first_name, ⁷​mdl_name, ⁸​last_name, ⁹​org_name
provider_enrollment(prov_type_desc = "PRACTITIONER - ENDOCRINOLOGY", state = "AK", gender = "F")
#> # A tibble: 4 × 11
#>   npi    pecos…¹ enrlm…² provi…³ provi…⁴ state…⁵ first…⁶ mdl_n…⁷ last_…⁸ org_n…⁹
#>   <chr>  <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 11846… 892109… I20040… 14-46   PRACTI… AK      JANICE  "A"     KOVAL   ""     
#> 2 17109… 094111… I20091… 14-46   PRACTI… AK      JEANNE  "R"     BONAR   ""     
#> 3 14977… 438567… I20151… 14-46   PRACTI… AK      CYDNEY  ""      FENTON  ""     
#> 4 10331… 953725… I20220… 14-46   PRACTI… AK      BRENDA  ""      GOODWIN ""     
#> # … with 1 more variable: gndr_sw <chr>, and abbreviated variable names
#> #   ¹​pecos_asct_cntl_id, ²​enrlmt_id, ³​provider_type_cd, ⁴​provider_type_desc,
#> #   ⁵​state_cd, ⁶​first_name, ⁷​mdl_name, ⁸​last_name, ⁹​org_name
provider_enrollment(prov_type_code = "14-41")
#> # A tibble: 49,047 × 11
#>    npi   pecos…¹ enrlm…² provi…³ provi…⁴ state…⁵ first…⁶ mdl_n…⁷ last_…⁸ org_n…⁹
#>    <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 1083… 599161… I20031… 14-41   PRACTI… NJ      ANNA    ""      MOY     ""     
#>  2 1467… 226434… I20031… 14-41   PRACTI… MN      DONALD  "W"     CALVIN  ""     
#>  3 1093… 387040… I20031… 14-41   PRACTI… PA      STEVEN  "JOHN"  RETO    ""     
#>  4 1326… 105222… I20031… 14-41   PRACTI… IN      CHRIST… "R"     PRATHER ""     
#>  5 1295… 185020… I20031… 14-41   PRACTI… OH      MARIA   "J"     FRAGOU… ""     
#>  6 1467… 084010… I20031… 14-41   PRACTI… PA      HARRY   "NICHO… HALSCH… ""     
#>  7 1649… 711383… I20031… 14-41   PRACTI… TX      SAMANT… "N"     NAIDOO  ""     
#>  8 1093… 498151… I20031… 14-41   PRACTI… MA      LEIGH   "S"     CHERVE… ""     
#>  9 1245… 256737… I20031… 14-41   PRACTI… MN      MARSHA… ""      TILLMAN ""     
#> 10 1285… 054717… I20031… 14-41   PRACTI… NJ      STEVEN  "A"     CIPPAR… ""     
#> # … with 49,037 more rows, 1 more variable: gndr_sw <chr>, and abbreviated
#> #   variable names ¹​pecos_asct_cntl_id, ²​enrlmt_id, ³​provider_type_cd,
#> #   ⁴​provider_type_desc, ⁵​state_cd, ⁶​first_name, ⁷​mdl_name, ⁸​last_name,
#> #   ⁹​org_name
provider_enrollment(gender = "9")
#> # A tibble: 3,528 × 11
#>    npi   pecos…¹ enrlm…² provi…³ provi…⁴ state…⁵ first…⁶ mdl_n…⁷ last_…⁸ org_n…⁹
#>    <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 1134… 286030… I20031… 14-41   PRACTI… TX      ROBERT  ""      YEAMAN  ""     
#>  2 1679… 599161… I20031… 14-25   PRACTI… AL      XIAOHUA ""      ZHOU    ""     
#>  3 1013… 852797… I20031… 14-29   PRACTI… PA      GERARD  ""      CRINER  ""     
#>  4 1801… 135525… I20031… 14-34   PRACTI… NJ      DEBRA   "L"     FROMER  ""     
#>  5 1083… 620474… I20031… 14-08   PRACTI… WA      NOEL    "M"     CHIA    ""     
#>  6 1518… 741686… I20031… 14-35   PRACTI… AK      LINDA   "L"     SCHERTZ ""     
#>  7 1205… 589069… I20031… 14-43   PRACTI… TN      MICHAEL "S"     STOVALL ""     
#>  8 1194… 347645… I20031… 14-08   PRACTI… NC      KEVAN   "R"     HANSEL  ""     
#>  9 1891… 731584… I20031… 14-13   PRACTI… OH      VIJAY   "R"     RAJAN   ""     
#> 10 1821… 711382… I20031… 14-50   PRACTI… TN      BENJAM… ""      MEEKS   ""     
#> # … with 3,518 more rows, 1 more variable: gndr_sw <chr>, and abbreviated
#> #   variable names ¹​pecos_asct_cntl_id, ²​enrlmt_id, ³​provider_type_cd,
#> #   ⁴​provider_type_desc, ⁵​state_cd, ⁶​first_name, ⁷​mdl_name, ⁸​last_name,
#> #   ⁹​org_name