Search the National Library of Medicine's ICD-10-CM API

provider_icd10 allows you to search the NLM's ICD-10-CM API by code or associated term.

Usage

provider_icd10(code = NULL, term = NULL, field = "both", limit = 500)

Arguments

code

All or part of an ICD-10-CM code

term

Associated term describing an ICD-10 code

field

options are "code" or "both"; default is "both"

limit

default is 700

Value

A tibble() containing the search results.

The NIH's Clinical Table Search Service API for ICD-10-CM

ICD-10-CM (International Classification of Diseases, 10th Revision, Clinical Modification) is a medical coding system for classifying diagnoses and reasons for visits in U.S. health care settings. Learn more about ICD-10-CM.

Current Version

ICD-10-CM 2023

Data Source

National Institute of Health/National Library of Medicine

Links

NIH NLM Clinical Table Service ICD-10-CM API

Examples

# Returns the seven ICD-10-CM codes beginning with "A15".
provider_icd10(code = "A15", limit = 10)
#> # A tibble: 7 × 2
#>   icd_10_cm_code icd_10_cm_term                              
#>   <chr>          <chr>                                       
#> 1 A15.0          Tuberculosis of lung                        
#> 2 A15.4          Tuberculosis of intrathoracic lymph nodes   
#> 3 A15.5          Tuberculosis of larynx, trachea and bronchus
#> 4 A15.6          Tuberculous pleurisy                        
#> 5 A15.7          Primary respiratory tuberculosis            
#> 6 A15.8          Other respiratory tuberculosis              
#> 7 A15.9          Respiratory tuberculosis unspecified        

# Returns the first 20 ICD-10-CM codes associated with tuberculosis.
provider_icd10(term = "tuber", limit = 20)
#> # A tibble: 20 × 2
#>    icd_10_cm_code icd_10_cm_term                                   
#>    <chr>          <chr>                                            
#>  1 A15.0          Tuberculosis of lung                             
#>  2 A15.4          Tuberculosis of intrathoracic lymph nodes        
#>  3 A15.5          Tuberculosis of larynx, trachea and bronchus     
#>  4 A15.6          Tuberculous pleurisy                             
#>  5 A15.7          Primary respiratory tuberculosis                 
#>  6 A15.8          Other respiratory tuberculosis                   
#>  7 A15.9          Respiratory tuberculosis unspecified             
#>  8 A17.0          Tuberculous meningitis                           
#>  9 A17.1          Meningeal tuberculoma                            
#> 10 A17.81         Tuberculoma of brain and spinal cord             
#> 11 A17.82         Tuberculous meningoencephalitis                  
#> 12 A17.83         Tuberculous neuritis                             
#> 13 A17.89         Other tuberculosis of nervous system             
#> 14 A17.9          Tuberculosis of nervous system, unspecified      
#> 15 A18.01         Tuberculosis of spine                            
#> 16 A18.02         Tuberculous arthritis of other joints            
#> 17 A18.03         Tuberculosis of other bones                      
#> 18 A18.09         Other musculoskeletal tuberculosis               
#> 19 A18.10         Tuberculosis of genitourinary system, unspecified
#> 20 A18.11         Tuberculosis of kidney and ureter                

# Returns the two ICD-10-CM codes associated with pleurisy.
provider_icd10(term = "pleurisy")
#> # A tibble: 2 × 2
#>   icd_10_cm_code icd_10_cm_term      
#>   <chr>          <chr>               
#> 1 R09.1          Pleurisy            
#> 2 A15.6          Tuberculous pleurisy

# Returns first 500 "Z" codes (Z00.00 - Z46.9)
provider_icd10(code = "z", field = "code")
#> # A tibble: 500 × 2
#>    icd_10_cm_code icd_10_cm_term                                                
#>    <chr>          <chr>                                                         
#>  1 Z00.00         Encounter for general adult medical examination without abnor…
#>  2 Z00.01         Encounter for general adult medical examination with abnormal…
#>  3 Z00.110        Health examination for newborn under 8 days old               
#>  4 Z00.111        Health examination for newborn 8 to 28 days old               
#>  5 Z00.121        Encounter for routine child health examination with abnormal …
#>  6 Z00.129        Encounter for routine child health examination without abnorm…
#>  7 Z00.2          Encounter for examination for period of rapid growth in child…
#>  8 Z00.3          Encounter for examination for adolescent development state    
#>  9 Z00.5          Encounter for examination of potential donor of organ and tis…
#> 10 Z00.6          Encounter for examination for normal comparison and control i…
#> # … with 490 more rows

# Must specify `field` as "code" or it will search term, for example:
# Returns codes with the letter "Z" in their associated term
provider_icd10(code = "z", limit = 10)
#> # A tibble: 10 × 2
#>    icd_10_cm_code icd_10_cm_term                                                
#>    <chr>          <chr>                                                         
#>  1 A28.8          Other specified zoonotic bacterial diseases, not elsewhere cl…
#>  2 A28.9          Zoonotic bacterial disease, unspecified                       
#>  3 A92.5          Zika virus disease                                            
#>  4 B02.0          Zoster encephalitis                                           
#>  5 B02.1          Zoster meningitis                                             
#>  6 B02.30         Zoster ocular disease, unspecified                            
#>  7 B02.31         Zoster conjunctivitis                                         
#>  8 B02.32         Zoster iridocyclitis                                          
#>  9 B02.33         Zoster keratitis                                              
#> 10 B02.34         Zoster scleritis                                              

# Returns the 95 codes beginning with "Z5-"
provider_icd10(code = "z5")
#> # A tibble: 95 × 2
#>    icd_10_cm_code icd_10_cm_term                                 
#>    <chr>          <chr>                                          
#>  1 Z51.0          Encounter for antineoplastic radiation therapy 
#>  2 Z51.11         Encounter for antineoplastic chemotherapy      
#>  3 Z51.12         Encounter for antineoplastic immunotherapy     
#>  4 Z51.5          Encounter for palliative care                  
#>  5 Z51.6          Encounter for desensitization to allergens     
#>  6 Z51.81         Encounter for therapeutic drug level monitoring
#>  7 Z51.89         Encounter for other specified aftercare        
#>  8 Z52.000        Unspecified donor, whole blood                 
#>  9 Z52.001        Unspecified donor, stem cells                  
#> 10 Z52.008        Unspecified donor, other blood                 
#> # … with 85 more rows