Search the Medicare Revalidation Due Date API

Information on revalidation due dates for Medicare providers.

Usage

provider_mrdd(
  npi = NULL,
  last = NULL,
  first = NULL,
  clean_names = TRUE,
  full = FALSE
)

Arguments

npi

10-digit National Provider Identifier (NPI)

last

Provider's last name

first

Provider's first name

clean_names

Clean column names with janitor's clean_names() function; default is TRUE.

full

If TRUE, downloads the entire dataset; default is FALSE.

Value

A tibble() containing the search results.

Medicare Revalidation Due Date API

The Revalidation Due Date List dataset contains revalidation due dates for Medicare providers who are due to revalidate in the following six months. If a provider's due date does not fall within the ensuing six months, the due date is marked 'TBD'. In addition the dataset also includes subfiles with reassignment information for a given provider as well as due date listings for clinics and group practices and their providers.

Data Update Frequency

Monthly

Data Source

Centers for Medicare & Medicaid Services

Links

Medicare Revalidation Due Date API

Examples

if (FALSE) {
provider_mrdd(npi = 1184699621)

provider_mrdd(last = "Byrd", first = "Eric")

# Unnamed List of NPIs
npi_list <- c(1003026055,
              1316405939,
              1720392988,
              1518184605,
              1922056829,
              1083879860)

npi_list |> purrr::map_dfr(provider_mrdd)

# Download Entire Dataset =================================================
provider_mrdd(full = TRUE)
}