Quality Payment Program Experience

quality_payment() allows the user access to information on participation and performance in the Merit-based Incentive Payment System (MIPS) and Advanced Alternative Payment Models (APMs) tracks.

Usage

quality_payment(
  year,
  npi = NULL,
  state = NULL,
  specialty = NULL,
  type = NULL,
  tidy = TRUE,
  nest = TRUE,
  eligibility = TRUE,
  ...
)

quality_payment_(year = qpp_years(), ...)

Arguments

year

< integer > // required Year data was reported, in YYYY format. Run qpp_years() to return a vector of the years currently available.

npi

< integer > 10-digit Individual National Provider Identifier assigned to the clinician when they enrolled in Medicare. Multiple rows for the same NPI indicate multiple TIN/NPI combinations.

state

< character > State or US territory code location of the TIN associated with the clinician.

specialty

< character > Specialty corresponding to the type of service that the clinician submitted most on their Medicare Part B claims for this TIN/NPI combination.

Nurse Practitioner

Internal Medicine

Physician Assistant

Family Practice

Emergency Medicine

Diagnostic Radiology

Anesthesiology

Neurology

Cardiology

type

< character > Participation type; level at which the performance data was collected, submitted or reported for the final score attributed to the clinician; drives most of the data returned.

"Group"

"Individual"

"MIPS APM"

tidy

< boolean > // default: TRUE Tidy output

nest

< boolean > // default: TRUE Nest status & measures

eligibility

< boolean > // default: TRUE Append results from quality_eligibility()

...

Pass arguments to quality_payment().

Value

A tibble containing the search results.

Quality Payment Program (QPP) Experience

The QPP dataset provides participation and performance information in the Merit-based Incentive Payment System (MIPS) during each performance year. They cover eligibility and participation, performance categories, final score and payment adjustments. The dataset provides additional details at the TIN/NPI level on what was published in the previous performance year.

Links

Quality Payment Program Experience

Update Frequency

Annually

Examples

if (FALSE) { # interactive()
quality_payment(year = 2020, npi = 1144544834)
}