Search the Medicare Revalidation Due Date List API

Information on revalidation due dates for Medicare providers. Medicare Providers must validate their enrollment record every three or five years. CMS sets every Provider’s Revalidation due date at the end of a month and posts the upcoming six to seven months of due dates online. A due date of ‘TBD’ means that CMS has not set the due date yet. These lists are refreshed every two months and two months’ worth of due dates are appended to the list

Usage

revalidation_date(
  npi = NULL,
  enroll_id = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  state = NULL,
  enroll_code = NULL,
  enroll_desc = NULL,
  specialty = NULL,
  tidy = TRUE
)

Arguments

npi

National Provider Identifier (NPI)

enroll_id

Enrollment ID

first

First name of individual provider

last

Last name of individual provider

organization

Legal business name of organizational provider

state

Enrollment state

enroll_code, enroll_desc

Provider enrollment type code or description:

1 = "Part A"

2 = "DME"

3 = "Non-DME Part B"

specialty

Enrollment specialty

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Details

The Revalidation Due Date List dataset contains revalidation due dates for Medicare providers who are due to revalidate in the following six months. If a provider's due date does not fall within the ensuing six months, the due date is marked 'TBD'. In addition the dataset also includes subfiles with reassignment information for a given provider as well as due date listings for clinics and group practices and their providers.

Links

Medicare Revalidation Due Date API

Update Frequency: Monthly

Examples

if (FALSE) { # interactive()
revalidation_date(enroll_id = "I20031110000070")
revalidation_date(enroll_id = "O20110620000324")
revalidation_date(state = "FL", enroll_code = 3, specialty = "General Practice")
}