Search the Medicare Revalidation Reassignment List API

Reassignments of Providers who are due for Revalidation.

Usage

revalidation_reassign(
  group_pac_id = NULL,
  group_enroll_id = NULL,
  group_bus_name = NULL,
  group_state = NULL,
  record_type = NULL,
  ind_enroll_id = NULL,
  ind_npi = NULL,
  ind_first = NULL,
  ind_last = NULL,
  ind_state = NULL,
  ind_specialty = NULL,
  clean_names = TRUE,
  lowercase = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

group_pac_id

PAC ID of provider who is receiving reassignment or is the employer

group_enroll_id

Enrollment ID of provider who is receiving reassignment or is the employer

group_bus_name

Legal business name of provider who is receiving reassignment or is the employer

group_state

Enrollment state of provider who is receiving reassignment or is the employer

record_type

Identifies whether the record is for a reassignment (Reassignment) or employment (Physician Assistant)

ind_enroll_id

Enrollment ID of provider who is reassigning their benefits or is an employee

ind_npi

NPI of provider who is reassigning their benefits or is an employee

ind_first

First name of provider who is reassigning their benefits or is an employee

ind_last

Last name of provider who is reassigning their benefits or is an employee

ind_state

Enrollment state of provider who is reassigning their benefits or is an employee

ind_specialty

Enrollment specialty of the provider who is reassigning their benefits or is an employee

clean_names

Clean column names with janitor's clean_names() function; default is TRUE.

lowercase

Convert column names to lowercase; default is TRUE.

Value

A tibble containing the search results.

Details

The Revalidation Reassignment List dataset provides information on reassignments of providers who are due for revalidation.

Links

Medicare Revalidation Reassignment List API

Note

Update Frequency: Monthly

Examples

revalidation_reassign(ind_enroll_id = "I20200929003184",
                   ind_npi = 1962026229,
                   ind_first = "Rashadda",
                   ind_last = "Wong",
                   ind_state = "CT",
                   ind_specialty = "Physician Assistant")
#> # A tibble: 3 × 16
#>   group_pac_id group_e…¹ group…² group…³ group…⁴ group…⁵ recor…⁶ indiv…⁷ indiv…⁸
#>          <dbl> <chr>     <chr>   <chr>   <chr>     <int> <chr>     <dbl> <chr>  
#> 1   4688988371 O2015081… 127197… CT      TBD          34 Physic…  2.77e9 I20200…
#> 2   3173866241 O2019051… Hartfo… CT      TBD        1676 Physic…  2.77e9 I20200…
#> 3   7618242777 O2017092… Hartfo… CT      TBD         262 Physic…  2.77e9 I20200…
#> # … with 7 more variables: individual_npi <int>, individual_first_name <chr>,
#> #   individual_last_name <chr>, individual_state_code <chr>,
#> #   individual_specialty_description <chr>, individual_due_date <chr>,
#> #   individual_total_employer_associations <int>, and abbreviated variable
#> #   names ¹​group_enrollment_id, ²​group_legal_business_name, ³​group_state_code,
#> #   ⁴​group_due_date, ⁵​group_reassignments_and_physician_assistants,
#> #   ⁶​record_type, ⁷​individual_pac_id, ⁸​individual_enrollment_id

revalidation_reassign(group_pac_id = 9436483807,
                   group_enroll_id = "O20190619002165",
                   group_bus_name = "1st Call Urgent Care",
                   group_state = "FL",
                   record_type = "Reassignment")
#> # A tibble: 1 × 16
#>   group_pac_id group_e…¹ group…² group…³ group…⁴ group…⁵ recor…⁶ indiv…⁷ indiv…⁸
#>          <dbl> <chr>     <chr>   <chr>   <chr>     <int> <chr>     <dbl> <chr>  
#> 1   9436483807 O2019061… 1st Ca… FL      TBD           1 Reassi…  5.29e9 I20190…
#> # … with 7 more variables: individual_npi <int>, individual_first_name <chr>,
#> #   individual_last_name <chr>, individual_state_code <chr>,
#> #   individual_specialty_description <chr>, individual_due_date <chr>,
#> #   individual_total_employer_associations <int>, and abbreviated variable
#> #   names ¹​group_enrollment_id, ²​group_legal_business_name, ³​group_state_code,
#> #   ⁴​group_due_date, ⁵​group_reassignments_and_physician_assistants,
#> #   ⁶​record_type, ⁷​individual_pac_id, ⁸​individual_enrollment_id