NUCC Provider Taxonomy Codeset

search_taxonomy() returns a tibble() of the current Health Care Provider Taxonomy code set

Usage

taxonomies(code = NULL, shape = c("wide", "long"), unnest = FALSE, ...)

Source

National Uniform Claim Committee

Arguments

code

<chr> vector of taxonomy codes

shape

<chr> shape of data to return, wide (default) or long

unnest

<lgl>unnest hierarchy column, default is FALSE

...

Empty dots

Value

A tibble containing the search results.

Details

Update Frequency: Biannually

Taxonomy Codes

The Health Care Provider Taxonomy code set is a collection of unique alphanumeric codes, ten characters in length. They contain no embedded logic.

The code set Levels are organized to allow for drilling down to the provider's most specific level of specialization, with three distinct "Levels":

Level I Provider Grouping

A major grouping of service(s) or occupation(s) of health care providers.

Examples: Allopathic & Osteopathic Physicians, Dental Providers, Hospitals

Level II Classification

A more specific service or occupation related to the Provider Grouping.

For example, the classification for Allopathic & Osteopathic Physicians is based upon the General Specialty Certificates as issued by the appropriate national boards.

The following boards will, however, have their general certificates appear as Level III Areas of specialization strictly due to display limitations of the code set for Boards that have multiple general certificates:

Medical Genetics, Preventive Medicine, Psychiatry & Neurology, Radiology, Surgery, Otolaryngology, Pathology

Level III Area of Specialization

A more specialized area of the Classification in which a provider chooses to practice or make services available.

For example, the area of specialization for Allopathic & Osteopathic Physicians is based upon the Subspecialty Certificates as issued by the appropriate national boards.

Categories (Level 0)

The code set includes three specialty categories:

Group (of Individuals):

Multi-Specialty

Single Specialty

Individuals:

Allopathic & Osteopathic Physicians

Behavioral Health and Social Service Providers

Chiropractic Providers

Dental Providers

Dietary and Nutritional Service Providers

Emergency Medical Service Providers

Eye and Vision Service Providers

Nursing Service Providers

Nursing Service Related Providers

Other Service Providers

Pharmacy Service Providers

Physician Assistants and Advanced Practice Nursing Providers

Podiatric Medicine and Surgery Service Providers

Respiratory, Developmental, Rehabilitative and Restorative Service Providers

Speech, Language and Hearing Service Providers

Student, Health Care

Technologists, Technicians, and Other Technical Service Providers

Non-Individuals:

Agencies

Ambulatory Health Care Facilities

Hospital Units

Hospitals

Laboratories

Managed Care Organizations

Nursing and Custodial Care Facilities

Other Service Providers

Residential Treatment Facilities

Respite Care Facilities

Suppliers

Transportation Services

Display Name

The display name is a combination of the code name and the Level in which the code is nested, which more precisely identifies the code.

For example, Addiction Medicine is a physician subspecialty in Anesthesiology, Family Medicine, Internal Medicine, Preventive Medicine, and Psychiatry & Neurology.

"Addiction Medicine" does not identify the specialty of the physician, but the display name of "Addiction Medicine (Internal Medicine) Physician" clearly does.

In another example, "Radiology" could be confused with several taxonomies, but "Radiology Chiropractor" more accurately specifies the provider.

Description

code: Provider Taxonomy Code

category: Indicates whether Taxonomy is Individual or Non-Individual, i.e., a group taxonomy

grouping: Level I, Provider Grouping

classification: Level II, Classification

specialization: Level III, Area of Specialization

display_name: Consumer-friendly taxonomy name, made of the code name and the Level in which the code is nested.

definition: Definition of Taxonomy

version: Three digit version of the code set. The first two digits indicate the year and the third digit indicates either the first release of the year ("0") or the second release of the year ("1").

release_date: Date the version of the code set was released

Examples

taxonomies(code = c("207K00000X", "193200000X"))
#> # A tibble: 2 × 10
#>   code    display_name section grouping classification specialization definition
#>   <chr>   <chr>        <chr>   <chr>    <chr>          <chr>          <chr>     
#> 1 193200… Multi-Speci… Indivi… Group    Multi-Special… NA             A busines…
#> 2 207K00… Allergy & I… Indivi… Allopat… Allergy & Imm… NA             An allerg…
#> # ℹ 3 more variables: notes <chr>, version <chr>, release_date <date>

taxonomies(code   = "207K00000X",
           shape  = "long",
           unnest = TRUE)
#> # A tibble: 3 × 6
#>   code       display_name                   level   description definition notes
#>   <chr>      <chr>                          <ord>   <chr>       <chr>      <chr>
#> 1 207K00000X Allergy & Immunology Physician I. Sec… Individual  An allerg… Sour…
#> 2 207K00000X Allergy & Immunology Physician II. Gr… Allopathic… An allerg… Sour…
#> 3 207K00000X Allergy & Immunology Physician III. C… Allergy & … An allerg… Sour…