Provider Facility Affiliations

affiliations() allows the user access to data concerning providers' facility affiliations

Usage

affiliations(
  npi = NULL,
  pac = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  facility_type = NULL,
  facility_ccn = NULL,
  parent_ccn = NULL,
  offset = 0L,
  tidy = TRUE,
  na.rm = TRUE,
  ...
)

Arguments

npi

<chr> Unique 10-digit National Provider Identifier number issued by CMS to US healthcare providers through NPPES.

pac

<int> Unique 10-digit Provider Associate-level Control ID (PAC), assigned to each individual or organization in PECOS. The PAC ID links all entity-level information (e.g., tax identification numbers and organizational names) and may be associated with multiple enrollment IDs if the individual or organization enrolled multiple times under different circumstances.

first, middle, last

<chr> Individual provider's first/middle/last name

facility_type

<chr> Type of facility, one of the following:

Hospital (hp)

Long-term care hospital (ltch)

Nursing home (nh)

Inpatient rehabilitation facility (irf)

Home health agency (hha)

Skilled nursing facility (snf)

Hospice (hs)

Dialysis facility (df)

facility_ccn

<chr> 6-digit CMS Certification Number of facility or unit within hospital where an individual provider provides service.

parent_ccn

<int> 6-digit CMS Certification Number of a sub-unit's primary hospital, should the provider provide services in said unit.

offset

<int> API pagination, default is 0L

tidy

<lgl> Tidy output; default is TRUE

na.rm

<lgl> Remove empty rows and columns; default is TRUE

...

Empty dots

Value

A tibble containing the search results.

Links

Physician Facility Affiliations

Certification Number (CCN) State Codes

Update Frequency: Monthly

Examples

if (FALSE) { # interactive()
affiliations(parent_ccn = 670055)
}