Provider Facility Affiliations

affiliations() allows the user access to data concerning providers' facility affiliations

Links:

Physician Facility Affiliations

Certification Number (CCN) State Codes

Update Frequency: Monthly

Usage

affiliations(
  npi = NULL,
  pac = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  facility_type = NULL,
  facility_ccn = NULL,
  parent_ccn = NULL,
  offset = 0L,
  tidy = TRUE,
  na.rm = TRUE
)

Arguments

npi

< integer > 10-digit Individual National Provider Identifier

pac

< integer > 10-digit Individual PECOS Associate Control ID

first, middle, last

< character > Individual Provider's name

facility_type

< character >

"Hospital" or "hp"

"Long-term care hospital" or "ltch"

"Nursing home" or "nh"

"Inpatient rehabilitation facility" or "irf"

"Home health agency" or "hha"

"Skilled nursing facility" or "snf"

"Hospice" or "hs"

"Dialysis facility" or "df"

facility_ccn

< character > 6-digit CMS Certification Number of facility or unit within hospital where an individual provider provides service

parent_ccn

< integer > 6-digit CMS Certification Number of a sub-unit's primary hospital, should the provider provide services in said unit

offset

< integer > // default: 0L API pagination

tidy

< boolean > // default: TRUE Tidy output

na.rm

< boolean > // default: TRUE Remove empty rows and columns

Value

A tibble with the columns:

Field	Description

npi	10-digit NPI

pac	10-digit individual PAC ID

first	Individual provider's first name

middle	Individual provider's middle name

last	Individual provider's last name

suffix	Individual provider's suffix

facility_type	Category of facility

facility_ccn	Facility's 6-digit CCN

parent_ccn	Primary facility's 6-digit CCN

Examples

if (FALSE) { # interactive()
affiliations(parent_ccn = 670055)
}