Clinical Laboratories

Clinical laboratories including demographics and the type of testing services the facility provides.

Usage

clia(
  facility_name = NULL,
  facility_ccn = NULL,
  parent_ccn = NULL,
  certificate = NULL,
  accreditation = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  compliant = NULL,
  active = NULL,
  count = FALSE,
  set = FALSE
)

Source

API: Provider of Services File - Clinical Laboratories

CMS: CLIA

CDC: CLIA

CMS: QCOR

CLIA: Certificates

CLIA: Certificate Fee Schedule

CLIA: Certification Guide

Arguments

facility_name

<chr> Provider/Laboratory name

facility_ccn

<chr> 10-digit CMS Certification Number

parent_ccn

<chr> 6-digit CMS Certification Number

certificate

<enum> CLIA certificate type (see Details):

"waiver" = Waiver

"ppm" = Provider-Performed Microscopy (PPM)

"registration" = Registration

"compliance" = Compliance

"accreditation" = Accreditation

accreditation

<enum> CLIA accrediting organization (see Details):

"a2la" = A2LA

"aabb" = AABB

"aoa" = AOA

"ashi" = ASHI-HLA

"cap" = CAP

"cola" = COLA

"jcaho" = JCAHO

city, state, zip

<chr> Lab city, state, zip

compliant

<lgl> Return only compliant or non-compliant labs

active

<lgl> Return only active labs

count

<lgl> Return the total row count

set

<lgl> Return the entire dataset

Value

A tibble containing the search results.

CLIA

CMS regulates all laboratory testing (except research) performed on humans in the U.S. through the Clinical Laboratory Improvement Amendments. In total, CLIA covers approximately 320,000 laboratory entities.

Although all clinical laboratories must be properly certified to receive Medicare or Medicaid payments, CLIA has no direct Medicare or Medicaid program responsibilities.

Certification

There are five CLIA certificate types, all of which are effective for a period of two years. They are as follows, in order of increasing complexity:

Waiver: Laboratory can only perform waived tests, those that are simple to perform. Although routine inspections are not conducted, 2% are visited annually to ensure quality. Does not waive the lab from all CLIA requirements.

Provider-Performed Microscopy (PPM): In addition to waived tests, a physician, mid-level practitioner or dentist can perform limited tests of moderate complexity requiring microscopic examination. No routine inspections are conducted.

Registration: Laboratory has applied for a Compliance or Accreditation certificate, enabling it to conduct moderate/high complexity testing until the survey is completed. Includes PPM and waived testing.

Compliance: Allows the laboratory to conduct moderate/high complexity testing and is issued after an inspection finds the lab to be in compliance with all applicable CLIA requirements. Includes PPM and waived testing.

Accreditation: Exactly the same as Compliance, except that the laboratory is accredited by one of the CMS-approved organizations:

A2LA: American Association for Laboratory Accreditation

AABB: Association for the Advancement of Blood & Biotherapies

AOA: American Osteopathic Association

ASHI-HLA: American Society for Histocompatibility & Immunogenetics

CAP: College of American Pathologists

COLA: Commission on Office Laboratory Accreditation

JCAHO: The Joint Commission

Examples

clia(count = TRUE)
#> Error in c("cli::cli_alert_info(c(\"{.strong {endpoint}} has \", \"{.strong {mark(x)}} \", ", "    \"{cli::qty(x)}row{?s}.\"))"): ! Could not evaluate cli `{}` expression: `mark(x)`.
#> Caused by error in `uuid_cms(self@end)`:
#> ! `endpoint` "eval" is invalid.

clia(compliant = FALSE, active = TRUE, count = TRUE)
#> Error in uuid_cms(self@end): `endpoint` "eval" is invalid.

clia(facility_ccn = provider:::cdc_labs$ccn)
#> Error in uuid_cms(self@end): `endpoint` "eval" is invalid.

clia(certificate = c("accreditation", "registration"), city = "Valdosta", state = "GA")
#> Error in uuid_cms(self@end): `endpoint` "eval" is invalid.

clia(accreditation = c("cap", "cola", "jcaho"))
#> Error in uuid_cms(self@end): `endpoint` "eval" is invalid.