Hospitals Enrolled in Medicare

hospitals() allows you to search for information on all hospitals currently enrolled in Medicare. Data returned includes the hospital's sub-group types, legal business name, doing-business-as name, organization type and address.

Links:

Hospital Enrollments

Update Frequency: Monthly

Usage

hospitals(
  npi = NULL,
  facility_ccn = NULL,
  enroll_id_org = NULL,
  enroll_state = NULL,
  specialty_code = NULL,
  pac_org = NULL,
  organization = NULL,
  doing_business_as = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  proprietary_nonprofit = NULL,
  multiple_npis = NULL,
  general = NULL,
  acute_care = NULL,
  alcohol_drug = NULL,
  childrens = NULL,
  long_term = NULL,
  psychiatric = NULL,
  rehabilitation = NULL,
  short_term = NULL,
  swing_bed = NULL,
  psych_unit = NULL,
  rehab_unit = NULL,
  specialty_hospital = NULL,
  other = NULL,
  reh_conversion = NULL,
  tidy = TRUE,
  pivot = TRUE,
  na.rm = TRUE
)

Arguments

npi

< integer > 10-digit national provider identifier

facility_ccn

< character > 6-digit CMS Certification Number of hospital

enroll_id_org

< character > 15-digit organizational provider Medicare enrollment identifier; begins with capital "O"

enroll_state

< character > Hospital’s enrollment state

specialty_code

< character > Enrollment specialty type code

pac_org

< integer > 10-digit organizational/group provider associate-level control identifier

organization

< character > Hospital’s legal business name

doing_business_as

< character > Hospital’s doing-business-as name

city

< character > City of the hospital’s practice location

state

< character > State of the hospital’s practice location

zip

< character > Zip code of the hospital’s practice location

proprietary_nonprofit

< character > "P" if hospital registered as proprietor with the IRS; "N" if registered as non-profit

multiple_npis

< boolean > Indicates hospital has more than one NPI

general

< boolean > Indicates hospital’s subgroup/unit is General

acute_care

< boolean > Indicates hospital’s subgroup/unit is Acute Care

alcohol_drug

< boolean > Indicates hospital’s subgroup/unit is Alcohol/Drug

childrens

< boolean > Indicates hospital’s subgroup/unit is Children's Hospital

long_term

< boolean > Indicates hospital’s subgroup/unit is Long-Term

psychiatric

< boolean > Indicates hospital’s subgroup/unit is Psychiatric

rehabilitation

< boolean > Indicates hospital’s subgroup/unit is Rehabilitation

short_term

< boolean > Indicates hospital’s subgroup/unit is Short-Term

swing_bed

< boolean > Indicates hospital’s subgroup/unit is Swing-Bed Approved

psych_unit

< boolean > Indicates hospital’s subgroup/unit is Psychiatric Unit

rehab_unit

< boolean > Indicates hospital’s subgroup/unit is Rehabilitation Unit

specialty_hospital

< boolean > Indicates hospital’s subgroup/unit is Specialty Hospital

other

< boolean > Indicates hospital’s subgroup/unit is not listed on the CMS form

reh_conversion

< boolean > Indicates a former Hospital or Critical Access Hospital that converted to a Rural Emergency Hospital

tidy

< boolean > // default: TRUE Tidy output

pivot

< boolean > // default: TRUE Pivot output

na.rm

< boolean > // default: TRUE Remove empty rows and columns

Value

A tibble containing the search results.

See also

Examples

hospitals(pac_org = 6103733050)
#> # A tibble: 2 × 18
#>   npi   organization pac_org enroll_id_org facility_ccn specialty_code specialty
#>   <chr> <chr>        <chr>   <chr>         <chr>        <chr>          <chr>    
#> 1 1164… ST CHARLES … 610373… O20020904000… 330246       00-09          PART A P…
#> 2 1225… ST CHARLES … 610373… O20090408000… 33T246       00-09          PART A P…
#> # ℹ 11 more variables: enroll_state <chr>, incorporation_date <date>,
#> #   incorporation_state <chr>, org_structure <chr>, address <chr>, city <chr>,
#> #   state <chr>, zip <int>, location_type <chr>, proprietary_nonprofit <chr>,
#> #   status <chr>

hospitals(state = "GA", reh_conversion = TRUE)
#> # A tibble: 1 × 20
#>   npi        organization   doing_business_as pac_org enroll_id_org facility_ccn
#>   <chr>      <chr>          <chr>             <chr>   <chr>         <chr>       
#> 1 1720098791 IRWIN COUNTY … PROGRESSIVE MEDI… 761895… O20230310002… 110779      
#> # ℹ 14 more variables: specialty_code <chr>, specialty <chr>,
#> #   enroll_state <chr>, incorporation_state <chr>, org_structure <chr>,
#> #   address <chr>, city <chr>, state <chr>, zip <int>, location_type <chr>,
#> #   associated_ccn <chr>, reh_conversion_date <date>,
#> #   proprietary_nonprofit <chr>, status <chr>