Clinical Laboratories

laboratories() allows you to search for information on clinical laboratories including demographics and the type of testing services the facility provides.

Clinical Laboratory Improvement Amendments (CLIA)

CMS regulates all laboratory testing (except research) performed on humans in the U.S. through the Clinical Laboratory Improvement Amendments (CLIA). In total, CLIA covers approximately 320,000 laboratory entities. The Division of Clinical Laboratory Improvement & Quality, within the Quality, Safety & Oversight Group, under the Center for Clinical Standards and Quality (CCSQ) has the responsibility for implementing the CLIA Program. Although all clinical laboratories must be properly certified to receive Medicare or Medicaid payments, CLIA has no direct Medicare or Medicaid program responsibilities.

CLIA Certificate types

There are five different CLIA certificate types which all are effective for a period of two years:

Certificate of Waiver Issued to a laboratory to perform only waived tests; does not waive the lab from all CLIA requirements. Waived tests are laboratory tests that are simple and easy to perform. Routine inspections are not conducted for waiver labs, though 2% are visited each year to ensure quality laboratory testing.

Certificate for Provider-Performed Microscopy Procedures (PPM) Issued to a laboratory in which a physician, midlevel practitioner or dentist performs limited tests that require microscopic examination. PPM tests are considered moderate complexity. Waived tests can also be performed under this certificate type. There are no routine inspections conducted for PPM labs.

Certificate of Registration Initially issued to a laboratory that has applied for a Certificate of Compliance or a Certificate of Accreditation, enabling the lab to conduct moderate/high complexity testing until the survey is performed and the laboratory found to be in compliance with the CLIA regulations. Includes PPM and waived testing.

Certificate of Compliance Allows the laboratory to conduct moderate/high complexity testing and is issued after an inspection finds the lab to be in compliance with all applicable CLIA requirements. Includes PPM and waived testing.

Certificate of Accreditation Exactly the same as the Certificate of Compliance, except that the laboratory must be accredited by one of the following CMS-approved accreditation organizations:

American Association for Laboratory Accreditation (AALA)

American Association of Blood Banks (AABB)

American Osteopathic Association (AOA)

American Society for Histocompatibility and Immunogenetics (ASHI)

College of American Pathologists (CAP)

Commission on Office Laboratory Accreditation (COLA)

the Joint Commission (JCAHO)

Links:

Provider of Services File - Clinical Laboratories

Update Frequency: Quarterly

Usage

laboratories(
  name = NULL,
  clia = NULL,
  certificate = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  tidy = TRUE,
  na.rm = TRUE,
  pivot = TRUE
)

Arguments

name

Provider or clinical laboratory's name

clia

10-character CLIA number

certificate

CLIA certificate type:

"waiver"

"compliance"

"accreditation"

"ppm" (Provider-Performed Microscopy Procedures)

"registration"

city

City

state

State

zip

Zip code

tidy

Tidy output; default is TRUE.

na.rm

Remove empty rows and columns; default is TRUE.

pivot

Pivot output; default is TRUE.

Value

A tibble containing the search results.

Examples

laboratories(clia = "11D0265516")
#> # A tibble: 1 × 22
#>   clia_number name     certificate effective_date expiration_date status poc_ind
#>   <chr>       <chr>    <chr>       <date>         <date>          <chr>  <lgl>  
#> 1 11D0265516  DANIEL … Compliance  1996-08-29     1998-08-04      In Co… TRUE   
#> # ℹ 15 more variables: termination_reason <chr>, type_of_action <chr>,
#> #   ownership_type <chr>, facility_type <chr>, director_affiliations <int>,
#> #   address <chr>, city <chr>, state <chr>, zip <chr>, phone <chr>,
#> #   orig_part_date <date>, application_date <date>, certification_date <date>,
#> #   mailed_date <date>, clia_class_current <chr>