Clinical Laboratories

laboratories() allows you to search for information on clinical laboratories including demographics and the type of testing services the facility provides.

Clinical Laboratory Improvement Amendments (CLIA)

CMS regulates all laboratory testing (except research) performed on humans in the U.S. through the Clinical Laboratory Improvement Amendments (CLIA). In total, CLIA covers approximately 320,000 laboratory entities. The Division of Clinical Laboratory Improvement & Quality, within the Quality, Safety & Oversight Group, under the Center for Clinical Standards and Quality (CCSQ) has the responsibility for implementing the CLIA Program. Although all clinical laboratories must be properly certified to receive Medicare or Medicaid payments, CLIA has no direct Medicare or Medicaid program responsibilities.

CLIA Certificates

There are five CLIA certificate types all of which are effective for a period of two years. They are as follows, in order of increasing complexity:

Certificate of Waiver: Issued to a laboratory to perform only waived tests; does not waive the lab from all CLIA requirements. Waived tests are laboratory tests that are simple to perform. Routine inspections are not conducted for waiver labs, although 2% are visited each year to ensure quality laboratory testing.

Certificate for Provider-Performed Microscopy Procedures (PPM): Issued to a laboratory in which a physician, midlevel practitioner or dentist performs limited tests that require microscopic examination. PPM tests are considered moderate complexity. Waived tests can also be performed under this certificate type. There are no routine inspections conducted for PPM labs.

Certificate of Registration: Initially issued to a laboratory that has applied for a Certificate of Compliance or Accreditation, enabling the lab to conduct moderate/high complexity testing until the survey is performed and the laboratory is found to be in CLIA compliance. Includes PPM and waived testing.

Certificate of Compliance: Allows the laboratory to conduct moderate/high complexity testing and is issued after an inspection finds the lab to be in compliance with all applicable CLIA requirements. Includes PPM and waived testing.

Certificate of Accreditation: Exactly the same as the Certificate of Compliance, except that the laboratory must be accredited by one of the following CMS-approved accreditation organizations:

American Association for Laboratory Accreditation (A2LA)

Association for the Advancement of Blood & Biotherapies (AABB)

American Osteopathic Association (AOA)

American Society for Histocompatibility and Immunogenetics (ASHI)

College of American Pathologists (CAP)

Commission on Office Laboratory Accreditation (COLA)

The Joint Commission (JCAHO)

Links:

Provider of Services File - Clinical Laboratories

Update Frequency: Quarterly

Usage

laboratories(
  name = NULL,
  clia = NULL,
  certificate = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  active = FALSE,
  tidy = TRUE,
  na.rm = TRUE,
  pivot = TRUE
)

Arguments

name

< character > Provider or clinical laboratory's name

clia

< character > 10-character CLIA number

certificate

< character > CLIA certificate type:

"waiver"

"ppm" (Provider-Performed Microscopy Procedures)

"registration"

"compliance"

"accreditation"

city

< character > City

state

< character > State

zip

< character > Zip code

active

< boolean > Search for Active providers only? dedault is FALSE

tidy

< boolean > Tidy output; default is TRUE

na.rm

< boolean > Remove empty rows and columns; default is TRUE

pivot

< boolean > Pivot output; default is TRUE

Value

A tibble containing the search results.

Examples

laboratories(clia = "11D0265516")
#> # A tibble: 1 × 23
#>   clia_number name            certificate effective_date expiration_date expired
#>   <chr>       <chr>           <chr>       <date>         <date>          <lgl>  
#> 1 11D0265516  DANIEL FELDMAN… Compliance  1996-08-29     1998-08-04      TRUE   
#> # ℹ 17 more variables: termination_reason <chr>, status <chr>, poc_ind <lgl>,
#> #   type_of_action <chr>, ownership_type <chr>, facility_type <chr>,
#> #   director_affiliations <int>, address <chr>, city <chr>, state <chr>,
#> #   zip <chr>, phone <chr>, orig_part_date <date>, application_date <date>,
#> #   certification_date <date>, mailed_date <date>, clia_class_current <chr>