Search the CMS Public Reporting of Missing Digital Contact Information API

Information on providers missing digital contact information in NPPES.

Usage

missing_information(npi = NULL, tidy = TRUE)

Source

Centers for Medicare & Medicaid Services

Arguments

npi

The provider’s National Provider Identifier

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Details

In the May 2020 CMS Interoperability and Patient Access final rule, CMS finalized the policy to publicly report the names and NPIs of those providers who do not have digital contact information included in the NPPES system (85 FR 25584). This data includes the NPI and provider name of providers and clinicians without digital contact information in NPPES.

Links

CMS Public Reporting of Missing Digital Contact Information API

Note

Update Frequency: Quarterly

Examples

missing_information(npi = 1134122013)
#> # A tibble: 1 × 3
#>   npi        last_name first_name
#>   <chr>      <chr>     <chr>     
#> 1 1134122013 Henry     Timothy   
missing_information(npi = 1386128379)
#> ✖ No results for npi: 1386128379