National Registry of All Health Care Providers

nppes() allows you to search the National Plan and Provider Enumeration System (NPPES) NPI Registry's public API, a free directory of all active National Provider Identifier (NPI) records.

National Provider Identifier
Healthcare providers acquire their unique 10-digit NPIs to identify themselves in a standard way throughout their industry. Once CMS supplies an NPI, they publish the parts of the NPI record that have public relevance, including the provider’s name, taxonomy and practice address.

Enumeration Type
Two categories of health care providers exist for NPI enumeration purposes:

Entity Type 1
Individual health care providers (including sole proprietors) may get an NPI as Entity Type 1. As a sole proprietor, they must apply for the NPI using their own SSN, not an Employer Identification Number (EIN) even if they have an EIN. As a sole proprietor, they may get only one NPI, just like any other individual. The following factors do not affect whether a sole proprietor is an Entity Type 1:

Number of different office locations

Whether they have employees

Whether the IRS issued them an EIN

An incorporated individual is a single health care provider who forms and conducts business under a corporation. A sole proprietor is not an incorporated individual because the sole proprietor didn't form a corporation. If you're a sole/solo practitioner, it doesn't necessarily mean you're a sole proprietor, and vice versa. If you're an individual health care provider who's incorporated, you may need to get an NPI for yourself (Entity Type 1) and an NPI for your corporation or LLC (Entity Type 2).

Entity Type 2
Organizational health care providers are group health care providers eligible for NPIs as Entity Type 2. Organization health care providers may have a single employee or thousands of employees. An example is an incorporated individual who is an organization's only employee. Some organization health care providers are made up of parts that work somewhat independently from their parent organization. These parts may offer different types of health care or offer health care in separate physical locations. These parts and their physical locations aren't themselves legal entities but are part of the organization health care provider (which is a legal entity). The NPI Final Rule refers to the parts and locations as sub-parts. An organization health care provider can get its sub-parts their own NPIs. If a sub-part conducts any HIPAA standard transactions on its own (separately from its parent), it must get its own NPI. Sub-part determination makes sure that entities within a covered organization are uniquely identified in HIPAA standard transactions they conduct with Medicare and other covered entities. For example, a hospital offers acute care, laboratory, pharmacy, and rehabilitation services. Each of these sub-parts may need its own NPI because each sends its own standard transactions to one or more health plans. Sub-part delegation doesn't affect Entity Type 1 health care providers. As individuals, these health care providers can't choose sub-parts and are not sub-parts.

Authorized Official
An appointed official (e.g., chief executive officer, chief financial officer, general partner, chairman of the board, or direct owner) to whom the organization has granted the legal authority to enroll it in the Medicare program, to make changes or updates to the organization's status in the Medicare program, and to commit the organization to fully abide by the statutes, regulations, and program instructions of the Medicare program.

Links

NPPES NPI Registry API Documentation

NPPES NPI Registry API Demo

Trailing Wildcard Entries Arguments that allow trailing wildcard entries are denoted in the parameter description with Trailing Wildcard Allowed. Wildcard entries require at least two characters to be entered, e.g. "jo*"

Update Frequency: Weekly

Usage

nppes(
  npi = NULL,
  entype = NULL,
  first = NULL,
  last = NULL,
  alias = TRUE,
  organization = NULL,
  name_type = NULL,
  taxonomy_desc = NULL,
  address_type = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  country = NULL,
  limit = 1200,
  skip = NULL,
  tidy = TRUE
)

Arguments

npi

10-digit National Provider Identifier (NPI)

entype

Entity/enumeration type. Cannot be the only criteria entered.

"I": Individual provider (NPI-1)

"O": Organizational provider (NPI-2)

first, last

Individual provider's first/last name. Trailing Wildcard Allowed

alias

"TRUE"/"FALSE". Applies to authorized officials and individual providers when not doing a wildcard search. When set to "TRUE", the results will include providers with similar first names.

organization

Healthcare organization's name. Many types of names (LBN, DBA, Former LBN, Other Name) may match. As such, the results might contain a name different from the one entered. Trailing Wildcard Allowed

name_type

Type of name the first/last arguments pertain to:

"AO": search Authorized Officials only

"Provider": search Individual Providers only (default)

taxonomy_desc

Provider's taxonomy description, e.g. "Pharmacist", "Pediatrics"

address_type

Address type of provider; options are:

"location" (Practice location)

"mailing"

"primary"

"secondary"

city

City associated with the provider's address. To search for a military address, enter either "APO" or "FPO".

state

State abbreviation associated with the provider's address. If this field is used, at least one other field, besides the entype and country, must be populated.

zip

Zip code associated with the provider's address. If a 5 digit zip is entered, it will be matched to any appropriate 9 digit (zip+4) codes in the data. Trailing Wildcard Allowed

country

Country abbreviation associated with the provider's address. Can be used as the only input criterion, as long as the value selected is not "US" (United States).

limit

Maximum number of results to return; default is 1200.

skip

Number of results to skip after those set in limit.

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Examples

if (FALSE) { # interactive()
nppes(npi = 1528060837, tidy = FALSE)
}