National Registry of All Health Care Providers

nppes() allows you to search the National Plan and Provider Enumeration System (NPPES) NPI Registry's public API, a free directory of all active National Provider Identifier (NPI) records.

National Provider Identifier
Healthcare providers acquire their unique 10-digit NPIs to identify themselves in a standard way throughout their industry. Once CMS supplies an NPI, they publish the parts of the NPI record that have public relevance, including the provider’s name, taxonomy and practice address.

Enumeration Type
Two categories of health care providers exist for NPI enumeration purposes:

Entity Type 1
Individual health care providers (including sole proprietors) may get an NPI as Entity Type 1. As a sole proprietor, they must apply for the NPI using their own SSN, not an Employer Identification Number (EIN) even if they have an EIN. As a sole proprietor, they may get only one NPI, just like any other individual. The following factors do not affect whether a sole proprietor is an Entity Type 1:

Number of different office locations

Whether they have employees

Whether the IRS issued them an EIN

An incorporated individual is a single health care provider who forms and conducts business under a corporation. A sole proprietor is not an incorporated individual because the sole proprietor didn't form a corporation. If you're a sole/solo practitioner, it doesn't necessarily mean you're a sole proprietor, and vice versa. If you're an individual health care provider who's incorporated, you may need to get an NPI for yourself (Entity Type 1) and an NPI for your corporation or LLC (Entity Type 2).

Entity Type 2
Organizational health care providers are group health care providers eligible for NPIs as Entity Type 2. Organization health care providers may have a single employee or thousands of employees. An example is an incorporated individual who is an organization's only employee. Some organization health care providers are made up of parts that work somewhat independently from their parent organization. These parts may offer different types of health care or offer health care in separate physical locations. These parts and their physical locations aren't themselves legal entities but are part of the organization health care provider (which is a legal entity). The NPI Final Rule refers to the parts and locations as sub-parts. An organization health care provider can get its sub-parts their own NPIs. If a sub-part conducts any HIPAA standard transactions on its own (separately from its parent), it must get its own NPI. Sub-part determination makes sure that entities within a covered organization are uniquely identified in HIPAA standard transactions they conduct with Medicare and other covered entities. For example, a hospital offers acute care, laboratory, pharmacy, and rehabilitation services. Each of these sub-parts may need its own NPI because each sends its own standard transactions to one or more health plans. Sub-part delegation doesn't affect Entity Type 1 health care providers. As individuals, these health care providers can't choose sub-parts and are not sub-parts.

Authorized Official
An appointed official (e.g., chief executive officer, chief financial officer, general partner, chairman of the board, or direct owner) to whom the organization has granted the legal authority to enroll it in the Medicare program, to make changes or updates to the organization's status in the Medicare program, and to commit the organization to fully abide by the statutes, regulations, and program instructions of the Medicare program.

Links

NPPES NPI Registry API Documentation

NPPES NPI Registry API Demo

Update Frequency: Weekly

Usage

nppes(
  npi = NULL,
  entype = NULL,
  first_name = NULL,
  last_name = NULL,
  org_name = NULL,
  purpose_name = NULL,
  taxonomy_desc = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  country = NULL,
  limit = 1200,
  skip = NULL,
  tidy = TRUE
)

Arguments

npi

10-digit National Provider Identifier (NPI).

entype

Entity type. Choices are either I for an Individual provider or O for an Organizational provider. Cannot be the only criteria entered.

first_name

Individual provider's first name. Trailing wildcard entries are permitted requiring at least two characters to be entered (e.g. jo* ).

last_name

Individual provider's last name. Trailing wildcard entries are permitted requiring at least two characters to be entered (e.g. jo* ).

org_name

Healthcare organization's name (NPI-2). Trailing wildcard entries are permitted requiring at least two characters to be entered. All types of Organization Names (LBN, DBA, Former LBN, Other Name) associated with an NPI are examined for matching contents, therefore, the results might contain an organization name different from the one entered in the Organization Name criterion.

purpose_name

Refers to whether the name information entered pertains to an Authorized Official's name or a Provider's name. When not specified, the results will search against a provider's first and last name. AO will only search against Authorized Official names. While PROVIDER will only search against Provider name. Valid values are: AO and Provider.

taxonomy_desc

Search for providers by their taxonomy by entering the taxonomy description.

city

City associated with the provider's address. To search for a Military Address, enter either APO or FPO.

state

State abbreviation associated with the provider's address. If this field is used, at least one other field, besides the entype and country, must be populated.

zip

The Postal Code associated with the provider's address identified in Address Purpose. If you enter a 5 digit postal code, it will match any appropriate 9 digit (zip+4) codes in the data. Trailing wildcard entries are permitted requiring at least two characters to be entered (e.g., 21*).

country

Country abbreviation associated with the provider's address. This field can be used as the only input criterion, as long as the value selected is not US (United States).

limit

Maximum number of results to return; default is 1200.

skip

Number of results to skip after those set in limit.

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Examples

nppes(npi = 1528060837, tidy = FALSE)
#> # A tibble: 1 × 21
#>   npi        entype first_name last_name middle_name credential sole_proprietor
#>   <chr>      <chr>  <chr>      <chr>     <chr>       <chr>      <chr>          
#> 1 1528060837 NPI-1  JOHN       SARGEANT  B           PT         NO             
#> # ℹ 14 more variables: gender <chr>, enumeration_date <chr>,
#> #   last_updated <chr>, status <chr>, name_prefix <chr>, name_suffix <chr>,
#> #   addresses <list>, taxonomy <list>, identifiers <list>, endpoints <lgl>,
#> #   practice_locations <lgl>, other_names <lgl>, created_epoch <chr>,
#> #   last_updated_epoch <chr>