Search CMS' Open Payments Program API

open_payments() allows you to search CMS' Open Payments Program API.

Usage

open_payments(
  covered_recipient_type = NULL,
  covered_recipient_profile_id = NULL,
  covered_recipient_npi = NULL,
  covered_recipient_first_name = NULL,
  covered_recipient_middle_name = NULL,
  covered_recipient_last_name = NULL,
  recipient_city = NULL,
  recipient_state = NULL,
  recipient_zip_code = NULL,
  applicable_manufacturer_or_applicable_gpo_making_payment_id = NULL,
  year = 2021,
  clean_names = TRUE,
  lowercase = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

covered_recipient_type

"Covered Recipient Physician"

covered_recipient_profile_id

Payment recipient's unique Open Payments ID

covered_recipient_npi

10-digit National Provider Identifier (NPI).

covered_recipient_first_name

Recipient's first name

covered_recipient_middle_name

Recipient's middle name

covered_recipient_last_name

Recipient's last name

recipient_city

City

recipient_state

State

recipient_zip_code

Zip code

applicable_manufacturer_or_applicable_gpo_making_payment_id

Paying entity's unique Open Payments ID

year

Reporting year, 2015-2021, default is 2021

clean_names

Clean column names with janitor's clean_names() function; default is TRUE.

lowercase

Convert column names to lowercase; default is TRUE.

Value

A tibble containing the search results.

Details

The Open Payments program is a national disclosure program that promotes a more transparent and accountable health care system. Open Payments houses a publicly accessible database of payments that reporting entities, including drug and medical device companies, make to covered recipients like physicians. Please note that CMS does not comment on what relationships may be beneficial or potential conflicts of interest. CMS publishes the data attested to by reporting entities. The data is open to individual interpretation.

Open Payments is a national transparency program that collects and publishes information about financial relationships between drug and medical device companies (referred to as "reporting entities") and certain health care providers (referred to as "covered recipients"). These relationships may involve payments to providers for things including but not limited to research, meals, travel, gifts or speaking fees.

The purpose of the program is to provide the public with a more transparent health care system. All information available on the Open Payments database is open to personal interpretation and if there are questions about what the data means, patients and their advocates should speak directly to the health care provider for a better understanding.

Links

What is the Open Payments Program?

OpenPaymentsData.cms.gov

2021 General Payment Data

2020 General Payment Data

2019 General Payment Data

2018 General Payment Data

2017 General Payment Data

2016 General Payment Data

2015 General Payment Data

Note

Update Frequency: Yearly

Examples

if (FALSE) {
open_payments(covered_recipient_npi = "1043218118")
}