Search the Medicare Opt Out Affidavits API

A list of practitioners who are currently opted out of Medicare.

Usage

opt_out(
  npi = NULL,
  first_name = NULL,
  last_name = NULL,
  specialty = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zipcode = NULL,
  order_and_refer = NULL,
  tidy = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

npi

National Provider Identifier (NPI) number of the Opt Out Provider

first_name

First Name of the Opt Out Provider

last_name

Last Name of the Opt Out Provider

specialty

Specialty of the Opt Out Provider

address

Provider's Street Address

city

Provider's City

state

Provider's State Abbreviation

zipcode

Provider's Zip Code

order_and_refer

Flag indicating whether the Provider is eligible to Order and Refer

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Details

The Opt Out Affidavits dataset provides information on providers who have decided not to participate in Medicare. It contains the provider's NPI, specialty, address, and effective dates.

Links

Medicare Opt Out Affidavits API

Note

Update Frequency: Monthly

Examples

opt_out(specialty = "Psychiatry", zipcode = "07626")
#> # A tibble: 3 × 13
#>   npi        first_name last_name  specialty  optout_start_date optout_end_date
#>   <chr>      <chr>      <chr>      <chr>      <date>            <date>         
#> 1 1235359019 Laura      Brenner    Psychiatry 2006-05-01        2024-05-01     
#> 2 1356579841 Ruchi      Nadella    Psychiatry 2016-03-30        2024-03-30     
#> 3 1508023920 Sireesha   Jathavedam Psychiatry 2018-01-22        2024-01-22     
#> # ℹ 7 more variables: optout_duration <Duration>, last_updated <date>,
#> #   order_and_refer <lgl>, address <chr>, city <chr>, state <chr>,
#> #   zipcode <chr>
opt_out(first = "David", last = "Smith")
#> # A tibble: 6 × 13
#>   npi        first_name last_name specialty    optout_start_date optout_end_date
#>   <chr>      <chr>      <chr>     <chr>        <date>            <date>         
#> 1 1649265760 David      Smith     Maxillofaci… 2012-06-25        2024-06-25     
#> 2 1750565958 David      Smith     Oral Surgery 2013-05-07        2023-05-07     
#> 3 1114058443 David      Smith     Psychiatry   2014-08-15        2024-08-15     
#> 4 1023002375 David      Smith     Obstetrics/… 2012-10-01        2024-10-01     
#> 5 1790853307 David      Smith     Oral Surgery 2015-03-26        2025-03-26     
#> 6 1659300077 David      Smith     Dentist      2015-05-26        2023-05-26     
#> # ℹ 7 more variables: optout_duration <Duration>, last_updated <date>,
#> #   order_and_refer <lgl>, address <chr>, city <chr>, state <chr>,
#> #   zipcode <chr>
opt_out(npi = 1114974490)
#> # A tibble: 1 × 13
#>   npi        first_name last_name specialty    optout_start_date optout_end_date
#>   <chr>      <chr>      <chr>     <chr>        <date>            <date>         
#> 1 1114974490 David      Altchek   Orthopedic … 2012-07-01        2024-07-01     
#> # ℹ 7 more variables: optout_duration <Duration>, last_updated <date>,
#> #   order_and_refer <lgl>, address <chr>, city <chr>, state <chr>,
#> #   zipcode <chr>
opt_out(city = "Los Angeles", address = "9201 W SUNSET BLVD")
#> # A tibble: 5 × 13
#>   npi        first_name last_name specialty    optout_start_date optout_end_date
#>   <chr>      <chr>      <chr>     <chr>        <date>            <date>         
#> 1 1649484007 Ned        Cowan     Psychiatry   2012-12-30        2024-12-30     
#> 2 1497783484 Geemee     Chung     Infectious … 2016-04-20        2024-04-20     
#> 3 1598825309 Genise     Fraiman   Internal Me… 2016-10-13        2024-10-13     
#> 4 1760448351 David      Salehani  Oral Surgery 2022-04-06        2024-04-06     
#> 5 1952408007 Colleen    Ryan      Family Prac… 2022-09-21        2024-09-21     
#> # ℹ 7 more variables: optout_duration <Duration>, last_updated <date>,
#> #   order_and_refer <lgl>, address <chr>, city <chr>, state <chr>,
#> #   zipcode <chr>
if (FALSE) {
opt_out(state = "NY", order_and_refer = FALSE)
# Returns empty list i.e., provider is not in the database
opt_out(npi = 1326011057)

# Example of possible data cleaning
psych <- opt_out(specialty = "Psychiatry")

psych |>
dplyr::mutate(last_updated = as.Date(parsedate::parse_date(last_updated)),
optout_effective_date = as.Date(parsedate::parse_date(optout_effective_date)),
optout_end_date = as.Date(parsedate::parse_date(optout_end_date))) |>
age_days(date, optout_end_date, colname = "days_until_end") |>
age_days(last_updated, date, colname = "days_since_update") |>
age_days(optout_effective_date, date, colname = "days_since_optout") |>
dplyr::mutate(zip_code = purrr::map_chr(zip_code, format_zipcode)) |>
dplyr::mutate(address_full = full_address(.data, "first_line_street_address",
"second_line_street_address", "city_name", "state_code","zip_code"))
}