Search the Medicare Opt Out Affidavits API

A list of practitioners who are currently opted out of Medicare.

Usage

opt_out(
  first = NULL,
  last = NULL,
  npi = NULL,
  specialty = NULL,
  date_start = NULL,
  date_end = NULL,
  address = NULL,
  city = NULL,
  state_abb = NULL,
  zip = NULL,
  eligible = NULL,
  clean_names = TRUE,
  lowercase = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

first

First Name of the Opt Out Provider

last

Last Name of the Opt Out Provider

npi

National Provider Identifier (NPI) number of the Opt Out Provider

specialty

Specialty of the Opt Out Provider

date_start

Date from which the Provider's Opt Out Status is effective

date_end

Date on which the Provider's Opt Out Status ends

address

Provider's Street Address

city

Provider's City

state_abb

Provider's State Abbreviation

zip

Provider's Zip Code

eligible

Flag indicating whether the Provider is eligible to Order and Refer

clean_names

Clean column names with janitor's clean_names() function; default is TRUE.

lowercase

Convert column names to lowercase; default is TRUE.

Value

A tibble containing the search results.

Details

The Opt Out Affidavits dataset provides information on providers who have decided not to participate in Medicare. It contains the provider's NPI, specialty, address, and effective dates.

Links

Medicare Opt Out Affidavits API

Note

Update Frequency: Monthly

Examples

opt_out(specialty = "Psychiatry", zip = "07626")
#> # A tibble: 3 × 14
#>   date       last_updated first_…¹ last_…² npi   speci…³ optou…⁴ optou…⁵ first…⁶
#>   <date>     <chr>        <chr>    <chr>   <chr> <chr>   <chr>   <chr>   <chr>  
#> 1 2023-01-21 11/15/2022   Laura    Brenner 1235… Psychi… 05/01/… 05/01/… 135 CO…
#> 2 2023-01-21 11/15/2022   Ruchi    Nadella 1356… Psychi… 03/30/… 03/30/… 135 CO…
#> 3 2023-01-21 11/15/2022   Sireesha Jathav… 1508… Psychi… 01/22/… 01/22/… 145 CO…
#> # … with 5 more variables: second_line_street_address <chr>, city_name <chr>,
#> #   state_code <chr>, zip_code <chr>, eligible_to_order_and_refer <lgl>, and
#> #   abbreviated variable names ¹​first_name, ²​last_name, ³​specialty,
#> #   ⁴​optout_effective_date, ⁵​optout_end_date, ⁶​first_line_street_address
opt_out(first = "David", last = "Smith")
#> # A tibble: 6 × 14
#>   date       last_updated first_…¹ last_…² npi   speci…³ optou…⁴ optou…⁵ first…⁶
#>   <date>     <chr>        <chr>    <chr>   <chr> <chr>   <chr>   <chr>   <chr>  
#> 1 2023-01-21 11/15/2022   David    Smith   1649… Maxill… 06/25/… 06/25/… 909 W …
#> 2 2023-01-21 11/15/2022   David    Smith   1750… Oral S… 05/07/… 05/07/… 1116 N…
#> 3 2023-01-21 11/15/2022   David    Smith   1114… Psychi… 08/15/… 08/15/… 3101 S…
#> 4 2023-01-21 11/15/2022   David    Smith   1023… Obstet… 10/01/… 10/01/… 755 N …
#> 5 2023-01-21 11/15/2022   David    Smith   1790… Oral S… 03/26/… 03/26/… 6800 M…
#> 6 2023-01-21 11/15/2022   David    Smith   1659… Dentist 05/26/… 05/26/… 3038 B…
#> # … with 5 more variables: second_line_street_address <chr>, city_name <chr>,
#> #   state_code <chr>, zip_code <chr>, eligible_to_order_and_refer <lgl>, and
#> #   abbreviated variable names ¹​first_name, ²​last_name, ³​specialty,
#> #   ⁴​optout_effective_date, ⁵​optout_end_date, ⁶​first_line_street_address
opt_out(npi = 1114974490)
#> # A tibble: 1 × 14
#>   date       last_updated first_…¹ last_…² npi   speci…³ optou…⁴ optou…⁵ first…⁶
#>   <date>     <chr>        <chr>    <chr>   <chr> <chr>   <chr>   <chr>   <chr>  
#> 1 2023-01-21 11/15/2022   David    Altchek 1114… Orthop… 07/01/… 07/01/… 535 EA…
#> # … with 5 more variables: second_line_street_address <chr>, city_name <chr>,
#> #   state_code <chr>, zip_code <chr>, eligible_to_order_and_refer <lgl>, and
#> #   abbreviated variable names ¹​first_name, ²​last_name, ³​specialty,
#> #   ⁴​optout_effective_date, ⁵​optout_end_date, ⁶​first_line_street_address
opt_out(state_abb = "NY", eligible = "N")
#> # A tibble: 621 × 14
#>    date       last_updated first…¹ last_…² npi   speci…³ optou…⁴ optou…⁵ first…⁶
#>    <date>     <chr>        <chr>   <chr>   <chr> <chr>   <chr>   <chr>   <chr>  
#>  1 2023-01-21 11/15/2022   Patrick Mcgrath 1932… Psychi… 06/01/… 06/01/… 100 PA…
#>  2 2023-01-21 11/15/2022   Stephen Donovan 1942… Psychi… 05/01/… 05/01/… 1070 P…
#>  3 2023-01-21 11/15/2022   William Golden  1477… Clinic… 07/01/… 07/01/… 4 RYDE…
#>  4 2023-01-21 11/15/2022   Ursula  Ofman   1427… Clinic… 10/01/… 10/01/… 347 FI…
#>  5 2023-01-21 11/15/2022   Lori    Davis   1902… Clinic… 07/01/… 07/01/… 80 FIF…
#>  6 2023-01-21 11/15/2022   Ronald  Hoffman 1679… Genera… 04/01/… 04/01/… 330 W …
#>  7 2023-01-21 11/15/2022   Linda   Penn    1033… Clinic… 05/21/… 05/21/… 98 REM…
#>  8 2023-01-21 11/15/2022   Paul    Geltner 1205… Clinic… 04/01/… 04/01/… 113 UN…
#>  9 2023-01-21 11/15/2022   Kathar… Strati… 1023… Psychi… 12/14/… 12/14/… 15 W 7…
#> 10 2023-01-21 11/15/2022   Marilyn Oswald  1790… Clinic… 07/19/… 07/19/… 257 CE…
#> # … with 611 more rows, 5 more variables: second_line_street_address <chr>,
#> #   city_name <chr>, state_code <chr>, zip_code <chr>,
#> #   eligible_to_order_and_refer <lgl>, and abbreviated variable names
#> #   ¹​first_name, ²​last_name, ³​specialty, ⁴​optout_effective_date,
#> #   ⁵​optout_end_date, ⁶​first_line_street_address
opt_out(date_start = "01/30/1998", date_end = "01/30/2024")
#> # A tibble: 1 × 14
#>   date       last_updated first_…¹ last_…² npi   speci…³ optou…⁴ optou…⁵ first…⁶
#>   <date>     <chr>        <chr>    <chr>   <chr> <chr>   <chr>   <chr>   <chr>  
#> 1 2023-01-21 11/15/2022   Jonathan Raines  1720… Psychi… 01/30/… 01/30/… 1629 M…
#> # … with 5 more variables: second_line_street_address <chr>, city_name <chr>,
#> #   state_code <chr>, zip_code <chr>, eligible_to_order_and_refer <lgl>, and
#> #   abbreviated variable names ¹​first_name, ²​last_name, ³​specialty,
#> #   ⁴​optout_effective_date, ⁵​optout_end_date, ⁶​first_line_street_address
opt_out(city = "Los Angeles", address = "9201 W SUNSET BLVD")
#> # A tibble: 5 × 14
#>   date       last_updated first_…¹ last_…² npi   speci…³ optou…⁴ optou…⁵ first…⁶
#>   <date>     <chr>        <chr>    <chr>   <chr> <chr>   <chr>   <chr>   <chr>  
#> 1 2023-01-21 11/15/2022   Ned      Cowan   1649… Psychi… 12/30/… 12/30/… 9201 W…
#> 2 2023-01-21 11/15/2022   Geemee   Chung   1497… Infect… 04/20/… 04/20/… 9201 W…
#> 3 2023-01-21 11/15/2022   Genise   Fraiman 1598… Intern… 10/13/… 10/13/… 9201 W…
#> 4 2023-01-21 11/15/2022   David    Saleha… 1760… Oral S… 04/06/… 04/06/… 9201 W…
#> 5 2023-01-21 11/15/2022   Colleen  Ryan    1952… Family… 09/21/… 09/21/… 9201 W…
#> # … with 5 more variables: second_line_street_address <chr>, city_name <chr>,
#> #   state_code <chr>, zip_code <chr>, eligible_to_order_and_refer <lgl>, and
#> #   abbreviated variable names ¹​first_name, ²​last_name, ³​specialty,
#> #   ⁴​optout_effective_date, ⁵​optout_end_date, ⁶​first_line_street_address
if (FALSE) {
# Returns empty list i.e., provider is not in the database
opt_out(npi = 1326011057)

# Example of possible data cleaning
psych <- opt_out(specialty = "Psychiatry")

psych |>
dplyr::mutate(last_updated = as.Date(parsedate::parse_date(last_updated)),
optout_effective_date = as.Date(parsedate::parse_date(optout_effective_date)),
optout_end_date = as.Date(parsedate::parse_date(optout_end_date))) |>
age_days(date, optout_end_date, colname = "days_until_end") |>
age_days(last_updated, date, colname = "days_since_update") |>
age_days(optout_effective_date, date, colname = "days_since_optout") |>
dplyr::mutate(zip_code = purrr::map_chr(zip_code, format_zipcode)) |>
dplyr::mutate(address_full = full_address(.data, "first_line_street_address",
"second_line_street_address", "city_name", "state_code","zip_code"))
}