Search the Medicare Opt Out Affidavits API

A list of practitioners who are currently opted out of Medicare.

Usage

opt_out(
  npi = NULL,
  first_name = NULL,
  last_name = NULL,
  specialty = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zipcode = NULL,
  order_and_refer = NULL,
  tidy = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

npi

National Provider Identifier (NPI) number of the Opt Out Provider

first_name

First Name of the Opt Out Provider

last_name

Last Name of the Opt Out Provider

specialty

Specialty of the Opt Out Provider

address

Provider's Street Address

city

Provider's City

state

Provider's State Abbreviation

zipcode

Provider's Zip Code

order_and_refer

Flag indicating whether the Provider is eligible to Order and Refer

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Details

The Opt Out Affidavits dataset provides information on providers who have decided not to participate in Medicare. It contains the provider's NPI, specialty, address, and effective dates.

Links

Medicare Opt Out Affidavits API

Note

Update Frequency: Monthly

Examples

if (FALSE) {
opt_out(specialty = "Psychiatry", zipcode = "07626")
opt_out(first = "David", last = "Smith")
opt_out(npi = 1114974490)
opt_out(city = "Los Angeles", address = "9201 W SUNSET BLVD")

opt_out(state = "NY", order_and_refer = FALSE)
# Returns empty list i.e., provider is not in the database
opt_out(npi = 1326011057)

# Example of possible data cleaning
psych <- opt_out(specialty = "Psychiatry")

psych |>
dplyr::mutate(last_updated = as.Date(parsedate::parse_date(last_updated)),
optout_effective_date = as.Date(parsedate::parse_date(optout_effective_date)),
optout_end_date = as.Date(parsedate::parse_date(optout_end_date))) |>
age_days(date, optout_end_date, colname = "days_until_end") |>
age_days(last_updated, date, colname = "days_since_update") |>
age_days(optout_effective_date, date, colname = "days_since_optout") |>
dplyr::mutate(zip_code = purrr::map_chr(zip_code, format_zipcode)) |>
dplyr::mutate(address_full = full_address(.data, "first_line_street_address",
"second_line_street_address", "city_name", "state_code","zip_code"))
}