Providers That Have Opted Out of Medicare

opt_out() allows you to search for information on providers who have decided not to participate in Medicare.

Opting Out

Providers who do not wish to enroll in the Medicare program may “opt-out” of Medicare. This means that neither the physician, nor the beneficiary submits the bill to Medicare for services rendered. Instead, the beneficiary pays the physician out-of-pocket and neither party is reimbursed by Medicare. A private contract is signed between the physician and the beneficiary that states that neither one can receive payment from Medicare for the services that were performed.

To opt out, a provider must:

Be of an eligible specialty type

Submit an opt-out affidavit to Medicare

Enter into a private contract with their Medicare patients, reflecting the agreement that they will pay out-of-pocket for services, and that no one will submit the bill to Medicare for reimbursement

Contact their Medicare Administrative Contractor (MAC) for instruction on information that should be included in their opt-out affidavit and private contract

Opt-Out Periods

Opt-out periods last for two years and cannot be terminated early unless the provider is opting out for the very first time and terminates the opt-out no later than 90 days after the opt-out period's effective date.

Opt-out statuses are effective for two years and automatically renew every two years. Providers that do not want to extend their opt-out status at the end of an opt-out period may cancel by notifying all Medicare contractors an affidavit was filed with at least 30 days prior to the start of the next opt-out period.

If a provider retires, surrenders their license, or no longer wants to participate in the Medicare program, they must officially withdraw within 90 days. DMEPOS suppliers must withdraw within 30 days.

Providers may NOT opt-out if they intend to be a Medicare Advantage (Part C) provider or furnish services covered by traditional Medicare fee-for-service (Part B).

Links:

Medicare Opt Out Affidavits API

Update Frequency: Monthly

Usage

opt_out(
  npi = NULL,
  first = NULL,
  last = NULL,
  specialty = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  order_refer = NULL,
  tidy = TRUE
)

Arguments

npi

10-digit National Provider Identifier

first, last

Provider's first/last name

specialty

Provider's specialty

address

Provider's address

city

Provider's city

state

Provider's state abbreviation

zip

Provider's zip code

order_refer

boolean indicating whether the provider is eligible to order and refer

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

See also

Examples

opt_out(first = "James",
        last = "Smith",
        specialty = "Nurse Practitioner",
        order_refer = TRUE)
#> # A tibble: 1 × 12
#>   npi        first last  specialty order_refer optout_start_date optout_end_date
#>   <chr>      <chr> <chr> <chr>     <lgl>       <date>            <date>         
#> 1 1043522824 James Smith Nurse Pr… TRUE        2019-07-01        2025-07-01     
#> # ℹ 5 more variables: last_updated <date>, address <chr>, city <chr>,
#> #   state <chr>, zip <chr>
if (FALSE) { # interactive()
# For providers that have opted out, but are eligible to order and refer,
# use `order_refer()` to look up their specific eligibility statuses
opt_out(first = "James",
        last = "Smith",
        specialty = "Nurse Practitioner",
        order_refer = TRUE) |>
        pull(npi) |>
        map(\(x) order_refer(npi = x)) |>
        list_rbind()
}