Ordering and Referring Eligibility for Medicare

order_refer() allows you to search for physicians and non-physician practitioners who are of a type/specialty that is legally eligible to order and refer to Part B (clinical laboratory and imaging), DME and Part A HHA claims in the Medicare program.

Ordering and Referring (or Certifying) Providers

Ordering providers can order non-physician services for patients. Referring providers can request items or services Medicare may reimburse on behalf of Medicare beneficiaries.

To qualify as an ordering and referring/certifying provider, a provider must:

have an individual NPI

be enrolled in Medicare in either an approved or opt-out status

be of an eligible specialty type

Providers currently enrolled in Medicare Part B can order and refer. A provider that does not want to bill Medicare for their services can still order and refer by opting out of Medicare or enrolling solely to order and refer. Medicare coverage will then apply when ordering or referring:

Durable Medical Equipment, Prosthetics, Orthotics, and Supplies (DMEPOS)

Clinical Laboratory Services

Imaging Services

Home Health Services

Links:

Medicare Order and Referring API

CMS.gov: Ordering & Certifying

Order and Referring Methodology

Update Frequency: Twice Weekly

Usage

order_refer(
  npi = NULL,
  first = NULL,
  last = NULL,
  partb = NULL,
  dme = NULL,
  hha = NULL,
  pmd = NULL,
  tidy = TRUE
)

Arguments

npi

10-digit National Provider Identifier

first, last

Provider's first/last name

partb

boolean; eligibility to order & refer to Medicare Part B

dme

boolean; eligibility to order Durable Medical Equipment

hha

boolean; eligibility to refer to a Home Health Agency

pmd

boolean; eligibility to order Power Mobility Devices

tidy

boolean; Tidy output; default is TRUE.

Value

A tibble object containing the search results.

See also

Examples

if (FALSE) { # interactive()
# Search by NPI
order_refer(npi = 1003026055)

# Search by name and/or filter for certain privileges
order_refer(last = "Smith", partb = FALSE, hha = TRUE)
}