Order and Referral Eligibility

order_refer() returns a provider's eligibility to order and refer within Medicare to:

Part B: Clinical Laboratory Services, Imaging Services

DME: Durable Medical Equipment, Prosthetics, Orthotics, & Supplies (DMEPOS)

Part A: Home Health Services

To be eligible, a provider must:

have an Individual NPI

be enrolled in Medicare in either an Approved or Opt-Out status

be of an Eligible Specialty type

Ordering Providers can order non-physician services for patients.

Referring (or Certifying) Providers can request items or services that Medicare may reimburse on behalf of its beneficiaries.

Opt-Out Providers: Providers who have opted out of Medicare may still order and refer. They can also enroll solely to order and refer.

Usage

order_refer(
  npi = NULL,
  first = NULL,
  last = NULL,
  partb = NULL,
  dme = NULL,
  hha = NULL,
  pmd = NULL,
  tidy = TRUE,
  pivot = TRUE
)

Arguments

npi

< integer > 10-digit national provider identifier

first, last

< character > Individual provider's first/last name

partb, dme, hha, pmd

< boolean > Whether a provider is eligible to order and refer to:

partb: Medicare Part B

dme: Durable Medical Equipment

hha: Home Health Agency

pmd: Power Mobility Devices

tidy

< boolean > // default: TRUE Tidy output

pivot

< boolean > // default: TRUE Pivot output

Value

A tibble with the columns:

Field	Description

npi	National Provider Identifier

first	Order and Referring Provider's First Name

last	Order and Referring Provider's Last Name

service	Services An Eligible Provider Can Order/Refer To

References

links:

Medicare Order and Referring API

CMS.gov: Ordering & Certifying

Order and Referring Methodology

Update Frequency: Twice Weekly

See also

Examples

order_refer(npi = 1003026055)
#> # A tibble: 4 × 4
#>   npi        first   last   service                  
#>   <chr>      <chr>   <chr>  <chr>                    
#> 1 1003026055 RADHIKA PHADKE Medicare Part B          
#> 2 1003026055 RADHIKA PHADKE Home Health Agency       
#> 3 1003026055 RADHIKA PHADKE Durable Medical Equipment
#> 4 1003026055 RADHIKA PHADKE Power Mobility Devices   

# Filter for certain privileges
order_refer(last = "Smith", partb = FALSE, hha = TRUE)
#> # A tibble: 9 × 4
#>   npi        first   last  service                  
#>   <chr>      <chr>   <chr> <chr>                    
#> 1 1386776425 GARRICK SMITH Home Health Agency       
#> 2 1386776425 GARRICK SMITH Durable Medical Equipment
#> 3 1386776425 GARRICK SMITH Power Mobility Devices   
#> 4 1215946959 LISA    SMITH Home Health Agency       
#> 5 1215946959 LISA    SMITH Durable Medical Equipment
#> 6 1215946959 LISA    SMITH Power Mobility Devices   
#> 7 1619457686 NAKIA   SMITH Home Health Agency       
#> 8 1619457686 NAKIA   SMITH Durable Medical Equipment
#> 9 1619457686 NAKIA   SMITH Power Mobility Devices