Search the Medicare Order and Referring API

All physicians and non-physician practitioners who are legally eligible to order and refer in the Medicare program and who have current enrollment records in Medicare.

Usage

order_refer(
  npi = NULL,
  last = NULL,
  first = NULL,
  partb = NULL,
  dme = NULL,
  hha = NULL,
  pmd = NULL,
  clean_names = TRUE,
  lowercase = TRUE
)

Arguments

npi

10-digit National Provider Identifier (NPI)

last

Provider's last name

first

Provider's first name

partb

logical

dme

logical

hha

logical

pmd

logical

clean_names

Clean column names with janitor's clean_names() function; default is TRUE.

lowercase

Convert column names to lowercase; default is TRUE.

Value

A tibble containing the search results.

Details

The Medicare Order and Referring dataset provides information on all physicians and non-physician practitioners, by their National Provider Identifier (NPI), who are of a type/specialty that is legally eligible to order and refer in the Medicare program and who have current enrollment records in Medicare.

Data Update Frequency

Weekly

Data Source

Centers for Medicare & Medicaid Services

Links

Medicare Order and Referring API

Examples

if (FALSE) {
order_refer(npi = 1003026055)

order_refer(last = "phadke",
            first = "radhika")

# Unnamed List of NPIs
npi_list <- c(1003026055,
              1316405939,
              1720392988,
              1518184605,
              1922056829,
              1083879860)

npi_list |>
purrr::map_dfr(order_refer)

# Data frame of NPIs
npi_df <- data.frame(npi = c(1003026055,
                             1316405939,
                             1720392988,
                             1518184605,
                             1922056829,
                             1083879860))
npi_df |>
tibble::deframe() |>
purrr::map_dfr(order_refer)
}