Search the Medicare Physician & Other Practitioners API by Provider

Information on services and procedures provided to Original Medicare (fee-for-service) Part B (Medical Insurance) beneficiaries by physicians and other healthcare professionals; aggregated by provider.

Usage

physician_by_provider(
  npi = NULL,
  last_org = NULL,
  first = NULL,
  cred = NULL,
  gender = NULL,
  type = NULL,
  city = NULL,
  state = NULL,
  fips = NULL,
  zip = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par_ind = NULL,
  year = 2020,
  clean_names = TRUE,
  lowercase = TRUE,
  nest = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

npi

National Provider Identifier (NPI) for the rendering provider on the claim. The provider NPI is the numeric identifier registered in NPPES.

last_org

Last name/Organization name of the provider. When the provider is registered in NPPES as an individual (entity type code I), this is the provider’s last name. When the provider is registered as an organization (entity type code O), this is the organization name.

first

An individual provider's (entity type code I) first name. An organization's (entity type code O) will be blank.

cred

An individual provider's (entity type code I) credentials. An organization's will be blank.

gender

An individual provider's gender. An organization's will be blank.

type

Type of entity reported in NPPES. An entity code of I identifies providers registered as individuals while an entity type code of O identifies providers registered as organizations.

city

The city where the provider is located, as reported in NPPES.

state

The state where the provider is located, as reported in NPPES.

fips

FIPS code for the rendering provider's state.

zip

The provider’s zip code, as reported in NPPES.

ruca

Rural-Urban Commuting Area Code (RUCA); a Census tract-based classification scheme that utilizes the standard Bureau of Census Urbanized Area and Urban Cluster definitions in combination with work commuting information to characterize all of the nation's Census tracts regarding their rural and urban status and relationships. The Referring Provider ZIP code was cross walked to the United States Department of Agriculture (USDA) 2010 Rural-Urban Commuting Area Codes.

country

The country where the provider is located, as reported in NPPES.

specialty

Derived from the provider specialty code reported on the claim. For providers that reported more than one specialty code on their claims, this is the specialty code associated with the largest number of services.

par_ind

Identifies whether the provider participates in Medicare and/or accepts assignment of Medicare allowed amounts. The value will be Y for any provider that had at least one claim identifying the provider as participating in Medicare or accepting assignment of Medicare allowed amounts within HCPCS code and place of service. A non-participating provider may elect to accept Medicare allowed amounts for some services and not accept Medicare allowed amounts for other services.

year

Year in YYYY format, between 2013-2020; default is 2020

clean_names

Clean column names with janitor's clean_names() function; default is TRUE.

lowercase

Convert column names to lowercase; default is TRUE.

nest

Nest related columns together; default is TRUE.

Value

A tibble containing the search results.

Details

The Provider dataset provides information on use, payments, submitted charges and beneficiary demographic and health characteristics organized by National Provider Identifier (NPI).

Links

Medicare Physician & Other Practitioners: by Provider API

Note

Update Frequency: Annually

Examples

physician_by_provider(npi = 1003000423,
                      year = 2020)
#> # A tibble: 1 × 11
#>    year rndrng_…¹ rndrng…² totals…³ drug_s…⁴ med_sr…⁵ bene_age bene_sex bene_r…⁶
#>   <dbl> <chr>     <list>   <list>   <list>   <list>   <list>   <list>   <list>  
#> 1  2020 10030004… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#> # … with 2 more variables: bene_status <list>, bene_cc <list>, and abbreviated
#> #   variable names ¹​rndrng_npi, ²​rndrng_prvdr, ³​totals_srvcs, ⁴​drug_srvcs,
#> #   ⁵​med_srvcs, ⁶​bene_race

physician_by_provider(type = "I",
                      city = "Hershey",
                      state = "PA",
                      fips = 42,
                      ruca =1,
                      gender = "F",
                      cred = "M.D.",
                      specialty = "Anesthesiology")
#> # A tibble: 19 × 11
#>     year rndrng…¹ rndrng…² totals…³ drug_s…⁴ med_sr…⁵ bene_age bene_sex bene_r…⁶
#>    <dbl> <chr>    <list>   <list>   <list>   <list>   <list>   <list>   <list>  
#>  1  2020 1023230… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#>  2  2020 1104213… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#>  3  2020 1114930… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#>  4  2020 1124240… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#>  5  2020 1154522… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#>  6  2020 1326249… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#>  7  2020 1376754… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#>  8  2020 1467763… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#>  9  2020 1518255… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#> 10  2020 1538506… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#> 11  2020 1558780… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#> 12  2020 1568489… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#> 13  2020 1578906… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#> 14  2020 1629289… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#> 15  2020 1710109… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#> 16  2020 1841466… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#> 17  2020 1942460… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#> 18  2020 1952796… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#> 19  2020 1972946… <tibble> <tibble> <tibble> <tibble> <tibble> <tibble> <tibble>
#> # … with 2 more variables: bene_status <list>, bene_cc <list>, and abbreviated
#> #   variable names ¹​rndrng_npi, ²​rndrng_prvdr, ³​totals_srvcs, ⁴​drug_srvcs,
#> #   ⁵​med_srvcs, ⁶​bene_race
if (FALSE) {
purrr::map_dfr(as.character(2013:2020), ~physician_by_provider(npi = 1003000126, year = .x))
}