Prescriber Utilization & Demographics by Year

prescribers() allows access to information on prescription drugs provided to Medicare beneficiaries enrolled in Part D (Prescription Drug Coverage), by physicians and other health care providers; aggregated by provider, drug and geography.

Usage

prescribers(
  year,
  type,
  npi = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  brand = NULL,
  generic = NULL,
  tidy = TRUE,
  na.rm = TRUE,
  ...
)

prescribers_(year = rx_years(), ...)

Arguments

year

< integer > // required Year data was reported, in YYYY format. Run rx_years() to return a vector of the years currently available.

type

< character > // required dataset to query, "provider", "drug", "geography"

npi

< integer > 10-digit national provider identifier

first, last, organization

< character > Individual/Organizational prescriber's name

credential

< character > Individual prescriber's credentials

gender

< character > Individual prescriber's gender; "F" (Female), "M" (Male)

entype

< character > Prescriber entity type; "I" (Individual), "O" (Organization)

city

< character > City where prescriber is located

state

< character > State where prescriber is located

zip

< character > Prescriber’s zip code

fips

< character > Prescriber's state's FIPS code

ruca

< character > Prescriber’s RUCA code

country

< character > Country where prescriber is located

specialty

< character > Prescriber specialty code reported on the largest number of claims submitted

brand

< character > Brand name (trademarked name) of the drug filled

generic

< character > USAN generic name of the drug filled (short version); A term referring to the chemical ingredient of a drug rather than the trademarked brand name under which the drug is sold

tidy

< boolean > // default: TRUE Tidy output

na.rm

< boolean > // default: TRUE Remove empty rows and columns

...

Pass arguments to prescribers().

By Provider

type ="provider":

The Provider dataset allows the user access to data such as services and procedures performed; charges submitted and payment received; and beneficiary demographic and health characteristics for providers treating Original Medicare (fee-for-service) Part B beneficiaries, aggregated by year.

By Provider and Drug

type ="service":

The Provider and Drug dataset is aggregated by:

Rendering provider's NPI

Healthcare Common Procedure Coding System (HCPCS) code

Place of Service (Facility or Non-facility)

There can be multiple records for a given NPI based on the number of distinct HCPCS codes that were billed and where the services were provided. Data have been aggregated based on the place of service because separate fee schedules apply depending on whether the place of service submitted on the claim is facility or non-facility.

By Geography and Drug

type ="geography":

The Geography and Drug dataset contains information on utilization, allowed amount, Medicare payment, and submitted charges organized nationally and state-wide by HCPCS code and place of service.

Links

Medicare Part D Prescribers: by Provider

Medicare Part D Prescribers: by Provider and Drug

Medicare Part D Prescribers: by Geography and Drug

Update Frequency: Annually

Examples

if (FALSE) { # interactive()
prescribers(year = 2020, type = 'provider', npi = 1003000423)
prescribers(year = 2021, type = 'provider', npi = 1003000126)
prescribers(year = 2019, type = 'drug', npi = 1003000126)

# Use the years helper function to retrieve results for every year:
rx_years() |>
map(\(x) prescribers(year = x, type = 'provider', npi = 1043477615)) |>
list_rbind()

# Parallelized version
prescribers_(type = 'provider', npi = 1043477615)
prescribers_(type = 'drug', npi = 1003000423)
}