Provider Enrollment in Medicare

provider_enrollment() allows you to access enrollment level data on individual and organizational providers that are actively approved to bill Medicare.

Usage

provider_enrollment(
  npi = NULL,
  pac_id = NULL,
  enroll_id = NULL,
  specialty_code = NULL,
  specialty = NULL,
  state = NULL,
  first_name = NULL,
  middle_name = NULL,
  last_name = NULL,
  organization_name = NULL,
  gender = NULL,
  tidy = TRUE
)

Arguments

npi

10-digit National Provider Identifier

pac_id

10-digit Provider associate level variable. Links all entity-level information and may be associated with multiple enrollment IDs if the individual or organization enrolled multiple times.

enroll_id

15-digit Provider enrollment ID. Assigned to each new provider enrollment application. Links all enrollment-level information (enrollment type, state, reassignment of benefits).

specialty_code

Enrollment primary specialty type code.

specialty

Enrollment specialty type description.

state

Enrollment state abbreviation. Since Providers enroll at the state level, a PAC ID can be associated with multiple enrollment IDs and states.

first_name

Individual provider first name

middle_name

Individual provider middle name

last_name

Individual provider last name

organization_name

Organizational provider name

gender

Individual provider gender:

F: Female

M: Male

9: Unknown or Organizational provider

tidy

Tidy output; default is TRUE.

Value

tibble containing the search results.

Links

Provider Enrollment API

Provider Enrollment Data Dictionary

Update Frequency

Quarterly

Examples

if (FALSE) { # interactive()
provider_enrollment(npi = 1417918293, specialty_code = "14-41")
provider_enrollment(first_name = "DEBRA", last_name = "FROMER")
provider_enrollment(organization_name = "Elizabethtown Community Hospital", state = "NY")
provider_enrollment(specialty = "PRACTITIONER - ENDOCRINOLOGY", state = "AK", gender = "F")
provider_enrollment(pac_id = 2860305554, gender = "9")
}