Search the CMS Quality Payment Program Experience API

Information on participation and performance in the Merit-based Incentive Payment System (MIPS) and Advanced Alternative Payment Models (APMs) tracks.

Usage

quality_payment(
  year,
  npi = NULL,
  state = NULL,
  specialty = NULL,
  part_type = NULL,
  tidy = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

year

year

npi

The NPI assigned to the clinician when they enrolled in Medicare. Multiple rows for the same NPI indicate multiple TIN/NPI combinations.

state

The State or United States (US) territory code location of the TIN associated with the clinician.

specialty

The specialty description is an identifier corresponding to the type of service that the clinician submitted most on their Medicare Part B claims for this TIN/NPI combination.

part_type

Indicates the level at which performance data was collected, submitted or reported for the final score attributed to the clinician. This information drives the data displayed for most of the remaining fields in this report.

tidy

Tidy output; default is TRUE.

Value

A tibble containing the search results.

Details

The Quality Payment Program (QPP) Experience dataset provides participation and performance information in the Merit-based Incentive Payment System (MIPS) during each performance year. They cover eligibility and participation, performance categories, and final score and payment adjustments. The dataset provides additional details at the TIN/NPI level on what was published in the previous performance year. You can sort the data by variables like clinician type, practice size, scores, and payment adjustments.

Links

Quality Payment Program Experience

Note

Update Frequency: Annually

Examples

if (FALSE) { # interactive()
quality_payment(year = 2020, state = "GA")
}