Reassignment of Benefits

reassignments() returns information about:

Individual providers who are reassigning benefits or are an employee of

Organizational/Group providers who are receiving reassignment of benefits from or are the employer of the individual provider

It provides information regarding the physician and the group practice they reassign their billing to, including individual employer association counts.

Usage

reassignments(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  first = NULL,
  last = NULL,
  state = NULL,
  specialty = NULL,
  organization = NULL,
  pac_org = NULL,
  enid_org = NULL,
  state_org = NULL,
  entry = NULL,
  tidy = TRUE,
  na.rm = TRUE,
  ...
)

Arguments

npi

< integer > Individual 10-digit National Provider Identifier

pac

< integer > Individual 10-digit PECOS Associate Control ID

enid

< character > Individual 15-digit Medicare Enrollment ID

first, last

< character > Individual Provider's name

state

< character > Individual Enrollment state abbreviation

specialty

< character > Individual Enrollment specialty

organization

< character > Organizational Legal business name

pac_org

< integer > Organizational 10-digit PECOS Associate Control ID

enid_org

< character > Organizational 15-digit Medicare Enrollment ID

state_org

< character > Organizational Enrollment state abbreviation

entry

< character > Entry type, reassignment ("R") or employment ("E")

tidy

< boolean > // default: TRUE Tidy output

na.rm

< boolean > // default: TRUE Remove empty rows and columns

...

Empty

Value

A tibble with the columns:

Field	Description

npi	Individual National Provider Identifier

pac	Individual PECOS Associate Control ID

enid	Individual Medicare Enrollment ID

first	Individual Provider's First Name

last	Individual Provider's Last Name

associations	Number of Organizations Individual Reassigns Benefits To

pac_org	Organization's PECOS Associate Control ID

enid_org	Organization's Medicare Enrollment ID

state_org	State Organization Enrolled in Medicare

reassignments	Number of Individuals the Organization Accepts Reassignment From

entry	Whether Entry is for Reassignment or Employment

Links

Medicare Revalidation Reassignment List API

Update Frequency: Monthly

Examples

if (FALSE) { # interactive()
reassignments(enid = "I20200929003184")

reassignments(pac = 9830437441)

reassignments(pac_org = 3173525888)
}