Search the Medicare Revalidation Due Date List API

Information on revalidation due dates for Medicare providers. Medicare Providers must validate their enrollment record every three or five years. CMS sets every Provider’s Revalidation due date at the end of a month and posts the upcoming six to seven months of due dates online. A due date of ‘TBD’ means that CMS has not set the due date yet. These lists are refreshed every two months and two months’ worth of due dates are appended to the list

Usage

revalidation_date(
  enroll_id = NULL,
  npi = NULL,
  first_name = NULL,
  last_name = NULL,
  org_name = NULL,
  state = NULL,
  type_code = NULL,
  prov_type = NULL,
  specialty = NULL,
  month = "Latest",
  clean_names = TRUE,
  lowercase = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

enroll_id

Enrollment ID

npi

National Provider Identifier (NPI)

first_name

First name of individual provider

last_name

Last name of individual provider

org_name

Legal business name of organizational provider

state

Enrollment state

type_code

Provider enrollment type code (1 if Part A; 2 if DME; 3 if Non-DME Part B)

prov_type

Provider type description

specialty

Enrollment specialty

month

dataset version, Latest is default; possible months are Feb, Apr, Jun, Jul, Aug, Sep, Oct, Nov,Dec

clean_names

Clean column names with janitor's clean_names() function; default is TRUE.

lowercase

Convert column names to lowercase; default is TRUE.

Value

A tibble containing the search results.

Details

The Revalidation Due Date List dataset contains revalidation due dates for Medicare providers who are due to revalidate in the following six months. If a provider's due date does not fall within the ensuing six months, the due date is marked 'TBD'. In addition the dataset also includes subfiles with reassignment information for a given provider as well as due date listings for clinics and group practices and their providers.

Links

Medicare Revalidation Due Date API

Note

Update Frequency: Monthly

Examples

revalidation_date(enroll_id = "I20031110000070",
                  npi = 1184699621)
#> # A tibble: 1 × 14
#>   month      enrollmen…¹ natio…² first…³ last_…⁴ organ…⁵ enrol…⁶ enrol…⁷ provi…⁸
#>   <date>     <chr>       <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 2022-12-21 I200311100… 118469… Eric    Byrd    ""      NC      3       Non-DM…
#> # … with 5 more variables: enrollment_specialty <chr>,
#> #   revalidation_due_date <chr>, adjusted_due_date <chr>,
#> #   individual_total_reassign_to <chr>, receiving_benefits_reassignment <int>,
#> #   and abbreviated variable names ¹​enrollment_id,
#> #   ²​national_provider_identifier, ³​first_name, ⁴​last_name, ⁵​organization_name,
#> #   ⁶​enrollment_state_code, ⁷​enrollment_type, ⁸​provider_type_text

revalidation_date(first_name = "Eric",
                  last_name = "Byrd")
#> # A tibble: 2 × 14
#>   month      enrollmen…¹ natio…² first…³ last_…⁴ organ…⁵ enrol…⁶ enrol…⁷ provi…⁸
#>   <date>     <chr>       <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 2022-12-21 I200311100… 118469… Eric    Byrd    ""      NC      3       Non-DM…
#> 2 2022-12-21 I200901280… 159879… Eric    Byrd    ""      SC      3       Non-DM…
#> # … with 5 more variables: enrollment_specialty <chr>,
#> #   revalidation_due_date <chr>, adjusted_due_date <chr>,
#> #   individual_total_reassign_to <chr>, receiving_benefits_reassignment <int>,
#> #   and abbreviated variable names ¹​enrollment_id,
#> #   ²​national_provider_identifier, ³​first_name, ⁴​last_name, ⁵​organization_name,
#> #   ⁶​enrollment_state_code, ⁷​enrollment_type, ⁸​provider_type_text

revalidation_date(state = "FL",
                  type_code = "3",
                  specialty = "General Practice")
#> # A tibble: 1,480 × 14
#>    month      enrollme…¹ natio…² first…³ last_…⁴ organ…⁵ enrol…⁶ enrol…⁷ provi…⁸
#>    <date>     <chr>      <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 2022-12-21 I20031104… 118469… John    Abt     ""      FL      3       Non-DM…
#>  2 2022-12-21 I20031105… 135643… Humber… Fernan… ""      FL      3       Non-DM…
#>  3 2022-12-21 I20031105… 146749… Jorge   Santan… ""      FL      3       Non-DM…
#>  4 2022-12-21 I20031111… 110487… Marian… De La … ""      FL      3       Non-DM…
#>  5 2022-12-21 I20031204… 143720… Mohamm… Sheikh  ""      FL      3       Non-DM…
#>  6 2022-12-21 I20031209… 141706… Richard Rizzo   ""      FL      3       Non-DM…
#>  7 2022-12-21 I20031211… 168964… Rene    Casano… ""      FL      3       Non-DM…
#>  8 2022-12-21 I20031211… 126558… Maria   Montea… ""      FL      3       Non-DM…
#>  9 2022-12-21 I20031215… 161994… Ricardo Doming… ""      FL      3       Non-DM…
#> 10 2022-12-21 I20031215… 103321… Arles   Perdomo ""      FL      3       Non-DM…
#> # … with 1,470 more rows, 5 more variables: enrollment_specialty <chr>,
#> #   revalidation_due_date <chr>, adjusted_due_date <chr>,
#> #   individual_total_reassign_to <chr>, receiving_benefits_reassignment <chr>,
#> #   and abbreviated variable names ¹​enrollment_id,
#> #   ²​national_provider_identifier, ³​first_name, ⁴​last_name, ⁵​organization_name,
#> #   ⁶​enrollment_state_code, ⁷​enrollment_type, ⁸​provider_type_text

revalidation_date(enroll_id = "O20110620000324",
                  org_name = "Lee Memorial Health System",
                  state = "FL",
                  prov_type = "DME",
                  type_code = "2")
#> # A tibble: 1 × 14
#>   month      enrollmen…¹ natio…² first…³ last_…⁴ organ…⁵ enrol…⁶ enrol…⁷ provi…⁸
#>   <date>     <chr>       <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 2022-12-21 O201106200… 180119… ""      ""      Lee Me… FL      2       DME    
#> # … with 5 more variables: enrollment_specialty <chr>,
#> #   revalidation_due_date <chr>, adjusted_due_date <chr>,
#> #   individual_total_reassign_to <chr>, receiving_benefits_reassignment <chr>,
#> #   and abbreviated variable names ¹​enrollment_id,
#> #   ²​national_provider_identifier, ³​first_name, ⁴​last_name, ⁵​organization_name,
#> #   ⁶​enrollment_state_code, ⁷​enrollment_type, ⁸​provider_type_text
if (FALSE) {
npi_list <- rep(c(1003026055,
                  1316405939,
                  1720392988,
                  1518184605),
                  each = 8)

months <- rep(c("Nov",
                "Oct",
                "Sep",
                "Aug",
                "Jul",
                "Jun",
                "Apr",
                "Feb"),
                times = 4)

purrr::map2_dfr(npi_list,
                months,
                ~revalidation_date(npi = .x,
                month = .y))
}