Search the Medicare Provider and Supplier Taxonomy Crosswalk API

A list of the type of providers and suppliers with the proper taxonomy code eligible for medicare programs.

Usage

taxonomy_crosswalk(
  taxonomy_code = NULL,
  taxonomy_desc = NULL,
  specialty_code = NULL,
  specialty_desc = NULL,
  clean_names = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

taxonomy_code

The taxonomy codes the providers use

taxonomy_desc

The description of the taxonomy that the providers use

specialty_code

Code that corresponds to the listed Medicare specialty

specialty_desc

Description of the Medicare Provider/Supplier Type

clean_names

Convert column names to snakecase; default is TRUE.

Value

A tibble containing the search results.

Details

The Medicare Provider and Supplier Taxonomy Crosswalk dataset lists the providers and suppliers eligible to enroll in Medicare programs with the proper healthcare provider taxonomy code. This data includes the Medicare speciality codes, if available, provider/supplier type description, taxonomy code, and the taxonomy description. This dataset is derived from information gathered from National Plan and Provider Enumerator System (NPPES) and Provider Enrollment, Chain and Ownership System (PECOS).

Links

Medicare Provider and Supplier Taxonomy Crosswalk API

Medicare Find Your Taxonomy Code

Note

Update Frequency: Weekly

Examples

taxonomy_crosswalk(specialty_code = "B4[14]")
#> # A tibble: 13 × 4
#>    taxonomy_code taxonomy_desc                     specialty_code specialty_desc
#>    <chr>         <chr>                             <chr>          <chr>         
#>  1 261QR0400X    Ambulatory Health Care Facilitie… B4[14]         Rehabilitatio…
#>  2 335U00000X    Suppliers/Organ Procurement Orga… B4[14]         Organ Procure…
#>  3 261QM0801X    Ambulatory Health Care Facilitie… B4[14]         Community Men…
#>  4 261QR0401X    Ambulatory Health Care Facilitie… B4[14]         Comprehensive…
#>  5 261QE0700X    Ambulatory Health Care Facilitie… B4[14]         End-Stage Ren…
#>  6 261QF0400X    Ambulatory Health Care Facilitie… B4[14]         Federally Qua…
#>  7 251G00000X    Agencies/Hospice Care Community … B4[14]         Hospice       
#>  8 291U00000X    Laboratories/Clinical Medical La… B4[14]         Histocompatib…
#>  9 291900000X    Laboratories/Military Clinical M… B4[14]         NA            
#> 10 261QR0400X    Ambulatory Health Care Facilitie… B4[14]         Outpatient Ph…
#> 11 315D00000X    Nursing & Custodial Care Facilit… B4[14]         Rehabilitatio…
#> 12 282J00000X    Hospitals/Religious Non-medical … B4[14]         Religious Non…
#> 13 261QR1300X    Ambulatory Health Care Facilitie… B4[14]         Rural Health …

taxonomy_crosswalk(specialty_desc = "Rehabilitation Agency")
#> # A tibble: 2 × 4
#>   taxonomy_code taxonomy_desc                      specialty_code specialty_desc
#>   <chr>         <chr>                              <chr>          <chr>         
#> 1 261QR0400X    Ambulatory Health Care Facilities… B4[14]         Rehabilitatio…
#> 2 315D00000X    Nursing & Custodial Care Faciliti… B4[14]         Rehabilitatio…

taxonomy_crosswalk(taxonomy_code = "2086S0102X")
#> # A tibble: 1 × 4
#>   taxonomy_code taxonomy_desc                      specialty_code specialty_desc
#>   <chr>         <chr>                              <chr>          <chr>         
#> 1 2086S0102X    Allopathic & Osteopathic Physicia… 2              Physician/Gen…

taxonomy_crosswalk(taxonomy_desc = "Agencies/Hospice Care Community Based")
#> # A tibble: 1 × 4
#>   taxonomy_code taxonomy_desc                      specialty_code specialty_desc
#>   <chr>         <chr>                              <chr>          <chr>         
#> 1 251G00000X    Agencies/Hospice Care Community B… B4[14]         Hospice