Utilization Statistics for Medicare Part B Providers

These functions allow the user access to information on services and procedures provided to Original Medicare (fee-for-service) Part B beneficiaries by physicians and other healthcare professionals; aggregated by provider, service and geography.

Usage

by_provider(
  year,
  npi = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par = NULL,
  tidy = TRUE,
  nest = TRUE
)

by_service(
  year,
  npi = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par = NULL,
  hcpcs_code = NULL,
  drug = NULL,
  pos = NULL,
  tidy = TRUE,
  rbcs = TRUE
)

by_geography(
  year,
  state = NULL,
  hcpcs_code = NULL,
  pos = NULL,
  level = NULL,
  fips = NULL,
  drug = NULL,
  tidy = TRUE,
  rbcs = TRUE
)

Arguments

year

< integer > // required Year data was reported, in YYYY format. Run helper function prac_years() to return a vector of the years currently available.

npi

< integer > 10-digit national provider identifier

first, last, organization

< character > Individual/Organizational provider's name

credential

< character > Individual provider's credentials

gender

< character > Individual provider's gender; "F" (Female), "M" (Male)

entype

< character > Provider entity type; "I" (Individual), "O" (Organization)

city

< character > City where provider is located

state

< character > State where provider is located

zip

< character > Provider’s zip code

fips

< character > Provider's state FIPS code

ruca

< character > Provider’s RUCA code

country

< character > Country where provider is located

specialty

< character > Provider specialty code reported on the largest number of claims submitted

par

< boolean > Identifies whether the provider participates in Medicare and/or accepts assignment of Medicare allowed amounts

tidy

< boolean > // default is TRUE Tidy output

nest

< boolean > // default: TRUE Nest hcpcs_detailed, demographics and conditions columns

hcpcs_code

< character > HCPCS code used to identify the specific medical service furnished by the provider

drug

< boolean > Identifies whether the HCPCS code is listed in the Medicare Part B Drug Average Sales Price (ASP) File

pos

< character > Identifies whether the Place of Service (POS) submitted on the claims is a:

Facility ("F"): Hospital, Skilled Nursing Facility, etc.

Non-facility ("O"): Office, Home, etc.

rbcs

< boolean > // default is TRUE Add Restructured BETOS Classifications to HCPCS codes

level

< character > Geographic level by which the data will be aggregated:

"State": Data is aggregated for each state

National: Data is aggregated across all states for a given HCPCS Code

Value

by_provider() returns a tibble with the columns:

Field	Description

year	Year data was reported

npi	10-digit national provider identifier

entity_type	Provider's entity/enumeration type

first	Individual provider's first name

middle	Individual provider's middle name

last	Individual provider's last name/Organization's name

gender	Individual provider's gender

credential	Individual provider's credential

specialty	Provider's specialty

address	Provider's street address

city	Provider's city

state	Provider's state

zip	Provider's zip code

fips	Provider's state's FIPS code

ruca	Provider's RUCA code

country	Provider's country

par	Indicates if provider participates in and/or accepts Medicare assignment

tot_hcpcs	Total number of unique HCPCS codes submitted

tot_benes	Total Medicare beneficiaries receiving services from the provider

tot_srvcs	Total services provided to beneficiaries by the provider

tot_charges	Total charges that the provider submitted for all services

tot_allowed	Total allowed amount for all services; the sum of Medicare reimbursement, deductible, coinsurance, etc.

tot_payment	Total amount Medicare paid, less any deductible and coinsurance

tot_std_pymt	Total amount Medicare paid, standardized to account for geographic differences

hcc_risk_avg	Average Hierarchical Condition Category (HCC) risk score of beneficiaries

hcpcs_detailed	Nested list columns drug & medical, offering further detail about the HCPCS codes submitted

demographics	Nested list column containing demographic data about the beneficiaries seen by the provider

conditions	Nested list column containing data about beneficiaries' chronic conditions

by_service() returns a tibble with the columns:

Field	Description

year	Year data was reported

npi	10-digit national provider identifier

level	Data aggregation level, will always be "Provider" here

entity_type	Provider's entity/enumeration type

first	Individual provider's first name

middle	Individual provider's middle name

last	Individual provider's last name/Organization's name

gender	Individual provider's gender

credential	Individual provider's credential

specialty	Provider's specialty

address	Provider's street address

city	Provider's city

state	Provider's state

zip	Provider's zip code

fips	Provider's state's FIPS code

ruca	Provider's RUCA code

country	Provider's country

par	Indicates if the provider participates in and/or accepts Medicare assignment

hcpcs_code	HCPCS code used to identify the specific medical service furnished by the provider

hcpcs_desc	Consumer Friendly Description of the HCPCS code

category	Restructured BETOS Classification System Category

subcategory	Restructured BETOS Classification System Subcategory

family	Restructured BETOS Classification System Family

procedure	Restructured BETOS Classification System Major Procedure Indicator

drug	Indicates if the HCPCS code is listed in the Medicare Part B Drug Average Sales Price (ASP) File

pos	Identifies the Place of Service (POS) submitted on the claim

tot_benes	Distinct number of Medicare beneficiaries for each HCPCS code/POS combination

tot_srvcs	Number of services provided for each HCPCS code/POS combination

tot_day	Number of distinct Medicare beneficiary/per day services for each HCPCS code/POS combination

avg_charge	Average charge that the provider submitted for the service

avg_allowed	Average of the Medicare allowed amount for the service

avg_payment	Average amount Medicare paid, less any deductible and coinsurance

avg_std_pymt	Average amount Medicare paid, standardized to account for geographic differences

by_geography() returns a tibble with the columns:

Field	Description

year	Year data was reported

level	Data aggregation level, either "National" or "State"

state	Provider's state

fips	Provider's state's FIPS code

hcpcs_code	HCPCS code used to identify the specific medical service furnished

hcpcs_desc	Consumer Friendly Description of the HCPCS code

category	Restructured BETOS Classification System Category

subcategory	Restructured BETOS Classification System Subcategory

family	Restructured BETOS Classification System Family

procedure	Restructured BETOS Classification System Major Procedure Indicator

drug	Indicates if the HCPCS code is listed in the Medicare Part B Drug Average Sales Price (ASP) File

pos	Identifies the Place of Service (POS) submitted on the claim

tot_provs	Number of providers that submitted each HCPCS code/POS combination

tot_benes	Distinct number of Medicare beneficiaries for each HCPCS code/POS combination

tot_srvcs	Number of services provided for each HCPCS code/POS combination

tot_day	Number of distinct Medicare beneficiary/per day services for each HCPCS code/POS combination

avg_charge	Average charge submitted for the service

avg_allowed	Average of the Medicare allowed amount for the service

avg_payment	Average amount Medicare paid, less any deductible and coinsurance

avg_std_pymt	Average amount Medicare paid, standardized to account for geographic differences

by_provider()

The Provider dataset allows the user access to data such as services and procedures performed; charges submitted and payment received; and beneficiary demographic and health characteristics for providers treating Original Medicare (fee-for-service) Part B beneficiaries, aggregated by year.

by_service()

The Provider and Service dataset is aggregated by:

Rendering provider's NPI

Healthcare Common Procedure Coding System (HCPCS) code

Place of Service (Facility or Non-facility)

There can be multiple records for a given NPI based on the number of distinct HCPCS codes that were billed and where the services were provided. Data have been aggregated based on the place of service because separate fee schedules apply depending on whether the place of service submitted on the claim is facility or non-facility.

by_geography()

The Geography and Service dataset contains information on utilization, allowed amount, Medicare payment, and submitted charges organized nationally and state-wide by HCPCS code and place of service.

Rural-Urban Commuting Area Codes (RUCA)

Metro Area Core

"1": Primary flow within Urbanized Area (UA)

"1.1": Secondary flow 30-50% to larger UA

Metro Area High Commuting

"2": Primary flow 30% or more to UA

"2.1": Secondary flow 30-50% to larger UA

Metro Area Low Commuting

"3": Primary flow 10-30% to UA

Micro Area Core

"4": Primary flow within large Urban Cluster (10k - 49k)

"4.1": Secondary flow 30-50% to UA

Micro High Commuting

"5": Primary flow 30% or more to large UC

"5.1": Secondary flow 30-50% to UA

Micro Low Commuting

"6": Primary flow 10-30% to large UC

Small Town Core

"7": Primary flow within small UC (2.5k - 9.9k)

"7.1": Secondary flow 30-50% to UA

"7.2": Secondary flow 30-50% to large UC

Small Town High Commuting

"8": Primary flow 30% or more to small UC

"8.1": Secondary flow 30-50% to UA

"8.2": Secondary flow 30-50% to large UC

Small Town Low Commuting

"9": Primary flow 10-30% to small UC

Rural Areas

"10": Primary flow to tract outside a UA or UC

"10.1": Secondary flow 30-50% to UA

"10.2": Secondary flow 30-50% to large UC

"10.3": Secondary flow 30-50% to small UC

"99": Zero population and no rural-urban identifier information

Links

Medicare Physician & Other Practitioners: by Provider API

Medicare Physician & Other Practitioners: by Provider and Service API

Medicare Physician & Other Practitioners: by Geography and Service API

Update Frequency: Annually

Examples

if (FALSE) { # interactive()
by_provider(year = 2020, npi = 1003000423)

by_service(year = 2019, npi = 1003000126)

by_geography(year = 2020, hcpcs_code = "0002A")

# Use the years helper function to retrieve results for every year:
prac_years() |>
map(\(x) by_provider(year = x, npi = 1043477615)) |>
list_rbind()
}