Provider Utilization & Demographics

These functions allow the user access to information on services and procedures provided to Original Medicare (fee-for-service) Part B beneficiaries by physicians and other healthcare professionals; aggregated by provider, service and geography.

Usage

utilization(
  year,
  type,
  npi = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par = NULL,
  level = NULL,
  hcpcs = NULL,
  drug = NULL,
  pos = NULL,
  tidy = TRUE,
  nest = TRUE,
  detailed = FALSE,
  rbcs = TRUE,
  na.rm = TRUE,
  ...
)

Arguments

year

< integer > // required Year data was reported, in YYYY format. Run util_years() to return a vector of the years currently available.

type

< character > // required dataset to query, "provider", "service", "geography"

npi

< integer > 10-digit national provider identifier

first, last, organization

< character > Individual/Organizational provider's name

credential

< character > Individual provider's credentials

gender

< character > Individual provider's gender; "F" (Female), "M" (Male)

entype

< character > Provider entity type; "I" (Individual), "O" (Organization)

city

< character > City where provider is located

state

< character > State where provider is located

zip

< character > Provider’s zip code

fips

< character > Provider's state's FIPS code

ruca

< character > Provider’s RUCA code

country

< character > Country where provider is located

specialty

< character > Provider specialty code reported on the largest number of claims submitted

par

< boolean > Identifies whether the provider participates in Medicare and/or accepts assignment of Medicare allowed amounts

level

< character > Geographic level by which the data will be aggregated:

"State": Data is aggregated for each state

National: Data is aggregated across all states for a given HCPCS Code

hcpcs

< character > HCPCS code used to identify the specific medical service furnished by the provider

drug

< boolean > Identifies whether the HCPCS code is listed in the Medicare Part B Drug Average Sales Price (ASP) File

pos

< character > Identifies whether the Place of Service (POS) submitted on the claims is a:

Facility ("F"): Hospital, Skilled Nursing Facility, etc.

Non-facility ("O"): Office, Home, etc.

tidy

< boolean > // default: TRUE Tidy output

nest

< boolean > // default: TRUE Nest demographics and conditions

detailed

< boolean > // default: FALSE Include detailed column

rbcs

< boolean > // default: TRUE Add Restructured BETOS Classifications to HCPCS codes

na.rm

< boolean > // default: TRUE Remove empty rows and columns

...

For future use.

type = "provider"

The Provider dataset allows the user access to data such as services and procedures performed; charges submitted and payment received; and beneficiary demographic and health characteristics for providers treating Original Medicare (fee-for-service) Part B beneficiaries, aggregated by year.

type = "service"

The Provider and Service dataset is aggregated by:

Rendering provider's NPI

Healthcare Common Procedure Coding System (HCPCS) code

Place of Service (Facility or Non-facility)

There can be multiple records for a given NPI based on the number of distinct HCPCS codes that were billed and where the services were provided. Data have been aggregated based on the place of service because separate fee schedules apply depending on whether the place of service submitted on the claim is facility or non-facility.

type = "geography"

The Geography and Service dataset contains information on utilization, allowed amount, Medicare payment, and submitted charges organized nationally and state-wide by HCPCS code and place of service.

Links

Medicare Physician & Other Practitioners: by Provider API

Medicare Physician & Other Practitioners: by Provider and Service API

Medicare Physician & Other Practitioners: by Geography and Service API

Update Frequency: Annually

Examples

if (FALSE) { # interactive()
utilization(year = 2020, type = "provider", npi = 1003000423)

utilization(year = 2019, type = "service", npi = 1003000126)

utilization(year = 2020, type = "geography", hcpcs = "0002A")

# Use the years helper function to retrieve results for every year:
util_years() |>
map(\(x) utilization(year = x, type = "provider", npi = 1043477615)) |>
list_rbind()

# Parallelized version
utilization_(npi = 1043477615, type = "provider")
utilization_(npi = 1043477615, type = "service")
utilization_(hcpcs = "0002A", type = "geography")
}